Crate bonsaidb

A programmable document database inspired by CouchDB written in Rust.

This crate provides a convenient way to access all functionality of BonsaiDb. The crates that are re-exported are:

bonsaidb-core: Common types and traits used when interacting with BonsaiDb.

bonsaidb-local: Local, file-based database implementation.

bonsaidb-server: Networked BonsaiDb server implementation.

bonsaidb-client: Client to access a BonsaiDb server.

Feature Flags

No feature flags are enabled by default in the bonsaidb crate. This is because in most Rust executables, you will only need a subset of the functionality. If you’d prefer to enable everything, you can use the full feature:

[dependencies]
bonsaidb = { version = "*", default-features = false, features = "full" }

full: Enables local-full, server-full, and client-full.

cli: Enables the bonsaidb executable.

Local databases only

[dependencies]
bonsaidb = { version = "*", default-features = false, features = "local-full" }

local-full: Enables local and local-cli

local: Enables the local module, which re-exports the crate bonsaidb-local.

local-cli: Enables the StructOpt structures for embedding database management commands into your own command-line interface.

BonsaiDb server

[dependencies]
bonsaidb = { version = "*", default-features = false, features = "server-full" }

server-full: Enables server and server-websockets

server: Enables the server module, which re-exports the crate bonsaidb-server.

server-websockets: Enables WebSocket support for bonsaidb-server.

Client for accessing a BonsaiDb server.

[dependencies]
bonsaidb = { version = "*", default-features = false, features = "client-full" }

client-full: Enables client, client-trusted-dns, and client-websockets

client: Enables the client module, which re-exports the crate bonsaidb-client.

client-trusted-dns: Enables using trust-dns for DNS resolution. If not enabled, all DNS resolution is done with the OS’s default name resolver.

client-websockets: Enables WebSocket support for bonsaidb-client.

Modules

cli

BonsaiDb command line tools.

client

Client for bonsaidb-server.

core

Core functionality and types for BonsaiDb.

keystorage

VaultKeyStorage implementors.

local

Local storage backend for BonsaiDb.

server

The BonsaiDb Server.