Module bonsaidb::cli

BonsaiDb command line tools.

Structs

Args

The command line interface for bonsaidb.

Enums

Command

All available command line commands.

NoSubcommand

A [Subcommand] implementor that has no options.

Traits

CommandLine

A command line interface that can be executed with either a remote or local connection to a server.

Functions

run

Runs the command-line interface with only the built-in commands, using configuration to launch a server if running a local command.