Crate bonsaidb::core

Core functionality and types for BonsaiDb.

Modules

actionable

Actionable provides the basic functionality needed to build an async-based API that has a flexible permissions system integrated.

admin

Database administration types and functionality.

api

Traits for tailoring a server.

arc_bytes

arc-bytes

async_trait

github crates-io docs-rs

circulate

crate version Live Build Status Documentation for main branch

connection

Types for interacting with BonsaiDb.

document

Types for interacting with Documents.

key

Key trait and related types.

keyvalue

Types for utilizing a lightweight atomic Key-Value store.

limits

Limits used within BonsaiDb.

networking

Types for implementing the BonsaiDb network protocol.

num_traits

Numeric traits for generic mathematics

ordered_varint

Provides variable-length signed and unsigned integer encoding that is byte-orderable.

permissions

Types for creating and validating permissions.

pubsub

Types for Publish/Subscribe (PubSub) messaging.

schema

Types for defining database schema.

test_util

Shared schemas and utilities used for unit testing.

transaction

Types for executing transactions.

transmog

Universal data serialization utilities for Rust.

transmog_pot

Transmog implementation of the Pot format.

Macros

assert_f64_eq

Compares two f64’s accounting for the epsilon.

define_async_connection_test_suite

Creates a test suite that tests methods available on AsyncConnection

define_async_kv_test_suite

Defines the KeyValue test suite

define_async_pubsub_test_suite

Expands into a suite of pubsub unit tests using the passed type as the test harness.

define_basic_mapped_view

Defines a non-unique view named $view_name for $collection with the mapping provided.

define_basic_unique_mapped_view

Defines an unique view named $view_name for $collection with the mapping provided.

define_blocking_connection_test_suite

Creates a test suite that tests methods available on AsyncConnection

define_blocking_kv_test_suite

Defines the KeyValue test suite

define_blocking_pubsub_test_suite

Expands into a suite of pubsub unit tests using the passed type as the test harness.

define_mapped_view

Defines a view using the mapping provided.

Enums

Error

an enumeration of errors that this crate can produce

Constants

ENCRYPTION_ENABLED

When true, encryption was enabled at build-time.

Traits

AnyError

A type that implements Error and is threadsafe.