Struct bonsaidb::core::connection::AsyncView

pub struct AsyncView<'a, Cn, V> where
    V: SerializedView,  {
    pub key: Option<QueryKey<<V as View>::Key>>,
    pub access_policy: AccessPolicy,
    pub sort: Sort,
    pub limit: Option<u32>,
    // some fields omitted
}

Parameters to query a schema::View.

The examples for this type use this view definition:

use bonsaidb_core::{
    define_basic_unique_mapped_view,
    document::{CollectionDocument, Emit},
    schema::{
        CollectionViewSchema, DefaultViewSerialization, Name, ReduceResult, View,
        ViewMapResult, ViewMappedValue,
    },
};

#[derive(Debug, Clone, View)]
#[view(collection = MyCollection, key = u32, value = f32, name = "scores-by-rank")]
pub struct ScoresByRank;

impl CollectionViewSchema for ScoresByRank {
    type View = Self;
    fn map(
        &self,
        document: CollectionDocument<<Self::View as View>::Collection>,
    ) -> ViewMapResult<Self::View> {
        document
            .header
            .emit_key_and_value(document.contents.rank, document.contents.score)
    }

    fn reduce(
        &self,
        mappings: &[ViewMappedValue<Self::View>],
        rereduce: bool,
    ) -> ReduceResult<Self::View> {
        if mappings.is_empty() {
            Ok(0.)
        } else {
            Ok(mappings.iter().map(|map| map.value).sum::<f32>() / mappings.len() as f32)
        }
    }
}

Fields
key: Option<QueryKey<<V as View>::Key>>

Key filtering criteria.

access_policy: AccessPolicy

The view’s data access policy. The default value is AccessPolicy::UpdateBefore.

sort: Sort

The sort order of the query.

limit: Option<u32>

The maximum number of results to return.

Implementations

impl<'a, Cn, V> AsyncView<'a, Cn, V> where
V: SerializedView,
Cn: AsyncConnection,

pub fn with_key(self, key: <V as View>::Key) -> AsyncView<'a, Cn, V>

Filters for entries in the view with key.

// score is an f32 in this example
for mapping in db.view::<ScoresByRank>().with_key(42).query().await? {
    assert_eq!(mapping.key, 42);
    println!("Rank {} has a score of {:3}", mapping.key, mapping.value);
}

pub fn with_keys<IntoIter>(self, keys: IntoIter) -> AsyncView<'a, Cn, V> where
IntoIter: IntoIterator<Item = <V as View>::Key>,

Filters for entries in the view with keys.

// score is an f32 in this example
for mapping in db
    .view::<ScoresByRank>()
    .with_keys([42, 43])
    .query()
    .await?
{
    println!("Rank {} has a score of {:3}", mapping.key, mapping.value);
}

pub fn with_key_range<R>(self, range: R) -> AsyncView<'a, Cn, V> where
R: Into<Range<<V as View>::Key>>,

Filters for entries in the view with the range keys.

// score is an f32 in this example
for mapping in db
    .view::<ScoresByRank>()
    .with_key_range(42..)
    .query()
    .await?
{
    assert!(mapping.key >= 42);
    println!("Rank {} has a score of {:3}", mapping.key, mapping.value);
}

pub fn with_key_prefix(self, prefix: <V as View>::Key) -> AsyncView<'a, Cn, V> where
<V as View>::Key: IntoPrefixRange,

Filters for entries in the view with keys that begin with prefix.

#[derive(View, Debug, Clone)]
#[view(name = "by-name", key = String, collection = MyCollection)]
struct ByName;

// score is an f32 in this example
for mapping in db
    .view::<ByName>()
    .with_key_prefix(String::from("a"))
    .query()
    .await?
{
    assert!(mapping.key.starts_with("a"));
    println!("{} in document {:?}", mapping.key, mapping.source);
}

pub fn with_access_policy(self, policy: AccessPolicy) -> AsyncView<'a, Cn, V>

Sets the access policy for queries.

// score is an f32 in this example
for mapping in db
    .view::<ScoresByRank>()
    .with_access_policy(AccessPolicy::UpdateAfter)
    .query()
    .await?
{
    println!("Rank {} has a score of {:3}", mapping.key, mapping.value);
}

pub fn ascending(self) -> AsyncView<'a, Cn, V>

Queries the view in ascending order. This is the default sorting behavior.

// score is an f32 in this example
for mapping in db.view::<ScoresByRank>().ascending().query().await? {
    println!("Rank {} has a score of {:3}", mapping.key, mapping.value);
}

pub fn descending(self) -> AsyncView<'a, Cn, V>

Queries the view in descending order.

// score is an f32 in this example
for mapping in db.view::<ScoresByRank>().descending().query().await? {
    println!("Rank {} has a score of {:3}", mapping.key, mapping.value);
}

pub fn limit(self, maximum_results: u32) -> AsyncView<'a, Cn, V>

Sets the maximum number of results to return.

// score is an f32 in this example
let mappings = db.view::<ScoresByRank>().limit(10).query().await?;
assert!(mappings.len() <= 10);

pub async fn query(
self
) -> Result<Vec<Map<<V as View>::Key, <V as View>::Value>, Global>, Error>

Executes the query and retrieves the results.

// score is an f32 in this example
for mapping in db.view::<ScoresByRank>().query().await? {
    println!("Rank {} has a score of {:3}", mapping.key, mapping.value);
}

pub async fn query_with_docs(
self
) -> Result<MappedDocuments<OwnedDocument, V>, Error>

Executes the query and retrieves the results with the associated Documents.

for mapping in &db
    .view::<ScoresByRank>()
    .with_key_range(42..=44)
    .query_with_docs()
    .await?
{
    println!(
        "Mapping from #{} with rank: {} and score: {}. Document bytes: {:?}",
        mapping.document.header.id, mapping.key, mapping.value, mapping.document.contents
    );
}

pub async fn query_with_collection_docs(
self
) -> Result<MappedDocuments<CollectionDocument<<V as View>::Collection>, V>, Error> where
<V as View>::Collection: SerializedCollection,
<<V as View>::Collection as SerializedCollection>::Contents: Debug,

Executes the query and retrieves the results with the associated CollectionDocuments.

for mapping in &db
    .view::<ScoresByRank>()
    .with_key_range(42..=44)
    .query_with_collection_docs()
    .await?
{
    println!(
        "Mapping from #{} with rank: {} and score: {}. Deserialized Contents: {:?}",
        mapping.document.header.id, mapping.key, mapping.value, mapping.document.contents
    );
}

pub async fn reduce(self) -> Result<<V as View>::Value, Error>

Executes a reduce over the results of the query

// score is an f32 in this example
let score = db.view::<ScoresByRank>().reduce().await?;
println!("Average score: {:3}", score);

pub async fn reduce_grouped(
self
) -> Result<Vec<MappedValue<<V as View>::Key, <V as View>::Value>, Global>, Error>

Executes a reduce over the results of the query, grouping by key.

// score is an f32 in this example
for mapping in db.view::<ScoresByRank>().reduce_grouped().await? {
    println!(
        "Rank {} has an average score of {:3}",
        mapping.key, mapping.value
    );
}

pub async fn delete_docs(self) -> Result<u64, Error>

Deletes all of the associated documents that match this view query.

db.view::<ScoresByRank>().delete_docs().await?;