Struct bonsaidb::core::connection::List

pub struct List<'a, Cn, Cl, PrimaryKey> where
    Cl: Collection,
    PrimaryKey: PartialEq<PrimaryKey> + ?Sized,
    <Cl as Collection>::PrimaryKey: Borrow<PrimaryKey>,
    <Cl as Collection>::PrimaryKey: PartialEq<PrimaryKey>,  { /* private fields */ }

Retrieves a list of documents from a collection. This structure also offers functions to customize the options for the operation.

Implementations

impl<'a, Cn, Cl, PrimaryKey> List<'a, Cn, Cl, PrimaryKey> where
Cl: Collection,
Cn: Connection,
PrimaryKey: 'a + for<'k> KeyEncoding<'k, <Cl as Collection>::PrimaryKey> + PartialEq<PrimaryKey> + ?Sized,
<Cl as Collection>::PrimaryKey: Borrow<PrimaryKey>,
<Cl as Collection>::PrimaryKey: PartialEq<PrimaryKey>,

pub fn ascending(self) -> List<'a, Cn, Cl, PrimaryKey>

Lists documents by id in ascending order.

pub fn descending(self) -> List<'a, Cn, Cl, PrimaryKey>

Lists documents by id in descending order.

pub fn limit(self, maximum_results: u32) -> List<'a, Cn, Cl, PrimaryKey>

Sets the maximum number of results to return.

pub fn count(self) -> Result<u64, Error>

Returns the number of documents contained within the range.

Order and limit are ignored if they were set.

println!(
    "Number of documents with id 42 or larger: {}",
    db.collection::<MyCollection>().list(&42..).count()?
);
println!(
    "Number of documents in MyCollection: {}",
    db.collection::<MyCollection>().all().count()?
);

pub fn headers(self) -> Result<Vec<Header, Global>, Error>

Returns the list of headers for documents contained within the range.

println!(
    "Headers with id 42 or larger: {:?}",
    db.collection::<MyCollection>().list(&42..).headers()?
);
println!(
    "Headers in MyCollection: {:?}",
    db.collection::<MyCollection>().all().headers()?
);

pub fn query(self) -> Result<Vec<OwnedDocument, Global>, Error>

Retrieves the matching documents.

for doc in db.collection::<MyCollection>().all().query()? {
    println!("Retrieved #{} with bytes {:?}", doc.header.id, doc.contents);
    let deserialized = MyCollection::document_contents(&doc)?;
    println!("Deserialized contents: {:?}", deserialized);
}