Struct bonsaidb::core::document::BorrowedDocument

pub struct BorrowedDocument<'a> {
    pub header: Header,
    pub contents: CowBytes<'a>,
}

Contains a serialized document in the database.

Fields
header: Header

The header of the document, which contains the id and Revision.

contents: CowBytes<'a>

The serialized bytes of the stored item.

Implementations

impl<'a> BorrowedDocument<'a>

pub fn new<Contents>(id: DocumentId, contents: Contents) -> BorrowedDocument<'a> where
Contents: Into<CowBytes<'a>>,

Returns a new instance with the id and content bytes.

pub fn with_contents<C, PrimaryKey>(
id: &PrimaryKey,
contents: &<C as SerializedCollection>::Contents
) -> Result<BorrowedDocument<'a>, Error> where
C: SerializedCollection,
PrimaryKey: for<'k> KeyEncoding<'k, <C as Collection>::PrimaryKey> + ?Sized,

Returns a new instance with contents, after serializing.

pub fn into_owned(self) -> OwnedDocument

Converts this document to an owned document.