Trait bonsaidb::core::document::Document

pub trait Document<'a>: Deref<Target = Header> + DerefMut + AsRef<[u8]> {
    type Bytes;
    fn new(id: u64, contents: impl Into<Self::Bytes>) -> Self;

    fn with_contents<S>(id: u64, contents: &S) -> Result<Self, Error>
    where
        S: SerializedCollection<Contents = S>;

    fn contents<D>(
        &self
    ) -> Result<<D as SerializedCollection>::Contents, Error>
    where
        D: SerializedCollection<Contents = D>;

    fn set_contents<S>(&mut self, contents: &S) -> Result<(), Error>
    where
        S: SerializedCollection<Contents = S>;

    fn create_new_revision(
        &self, 
        contents: impl Into<Self::Bytes>
    ) -> Option<Self>;
}

Common interface of a document in BonsaiDb.

Associated Types

type Bytes

The bytes type used in the interface.

Required methods

fn new(id: u64, contents: impl Into<Self::Bytes>) -> Self

Creates a new document with contents.

fn with_contents<S>(id: u64, contents: &S) -> Result<Self, Error> where
S: SerializedCollection<Contents = S>,

Creates a new document with serialized bytes from contents.

fn contents<D>(&self) -> Result<<D as SerializedCollection>::Contents, Error> where
D: SerializedCollection<Contents = D>,

Retrieves contents through deserialization into the type D.

fn set_contents<S>(&mut self, contents: &S) -> Result<(), Error> where
S: SerializedCollection<Contents = S>,

Serializes and stores contents into this document.

fn create_new_revision(&self, contents: impl Into<Self::Bytes>) -> Option<Self>

Creates a new revision.

WARNING: This normally should not be used outside of implementing a backend for BonsaiDb. To update a document, use set_contents() and send the document with the existing Revision information.