Enum bonsaidb::core::Error

pub enum Error {

    SchemaMismatch {
        database_name: String,
        schema: SchemaName,
        stored_schema: SchemaName,
    },
    SchemaAlreadyRegistered(SchemaName),
    SchemaNotRegistered(SchemaName),
    ViewAlreadyRegistered(ViewName),
    InvalidDatabaseName(String),
    DatabaseNotFound(String),
    ViewNotFound,
    CollectionNotFound,
    ApiNotFound(ApiName),
    DatabaseNameAlreadyTaken(String),
    Networking(Error),
    CollectionAlreadyDefined,
    DocumentNotFound(CollectionName, Box<DocumentId, Global>),
    DocumentIdTooLong,
    DocumentConflict(CollectionName, Box<Header, Global>),
    UniqueKeyViolation {
        view: ViewName,
        conflicting_document: Box<Header, Global>,
        existing_document: Box<Header, Global>,
    },
    DocumentPush(CollectionName, NextValueError),
    InvalidName(InvalidNameError),
    PermissionDenied(PermissionDenied),
    Password(String),
    UserNotFound,
    InvalidUnicode(String),
    InvalidCredentials,
    ReduceUnimplemented,
    NotANumber,
    Time(TimeError),
    Other {
        origin: String,
        error: String,
    },

}

an enumeration of errors that this crate can produce

Variants

SchemaMismatch

Fields

database_name: String

The name of the database being accessed.

schema: SchemaName

The schema provided for the database.

stored_schema: SchemaName

The schema stored for the database.

The database named database_name was created with a different schema (stored_schema) than provided (schema).

SchemaAlreadyRegistered(SchemaName)

The SchemaName returned has already been registered.

SchemaNotRegistered(SchemaName)

The SchemaName requested was not registered.

ViewAlreadyRegistered(ViewName)

The ViewName returned has already been registered.

InvalidDatabaseName(String)

An invalid database name was specified. See StorageConnection::create_database() for database name requirements.

DatabaseNotFound(String)

The database name given was not found.

ViewNotFound

The view was not found.

CollectionNotFound

The collection was not found.

ApiNotFound(ApiName)

The api invoked was not found.

DatabaseNameAlreadyTaken(String)

The database name already exists.

Networking(Error)

An error occurred from networking.

CollectionAlreadyDefined

A Collection being added already exists. This can be caused by a collection name not being unique.

DocumentNotFound(CollectionName, Box<DocumentId, Global>)

An attempt to update a document that doesnโt exist.

DocumentIdTooLong

A value provided as a DocumentId exceeded DocumentId::MAX_LENGTH.

DocumentConflict(CollectionName, Box<Header, Global>)

When updating a document, if a situation is detected where the contents have changed on the server since the Revision provided, a Conflict error will be returned.

UniqueKeyViolation

Fields

view: ViewName

The name of the view that the unique key violation occurred.

conflicting_document: Box<Header, Global>

The document that caused the violation.

existing_document: Box<Header, Global>

The document that already uses the same key.

When saving a document in a collection with unique views, a document emits a key that is already emitted by an existing ocument, this error is returned.

DocumentPush(CollectionName, NextValueError)

When pushing a document, an error occurred while generating the next unique id.

InvalidName(InvalidNameError)

An invalid name was specified during schema creation.

PermissionDenied(PermissionDenied)

Permission was denied.

Password(String)

An internal error handling passwords was encountered.

UserNotFound

The user specified was not found. This will not be returned in response to an invalid username being used during login. It will be returned in other APIs that operate upon users.

InvalidUnicode(String)

An error occurred converting from bytes to Utf-8.

InvalidCredentials

The credentials specified are not valid.

ReduceUnimplemented

Returned when the a viewโs reduce() function is unimplemented.

NotANumber

A floating point operation yielded Not a Number.

Time(TimeError)

An error while operating with a time

Other

Fields

origin: String

The origin of the error.

error: String

The error message.

An error from another crate.

Implementations

impl Error

pub fn other(origin: impl Display, error: impl Display) -> Error

Returns an instance of Self::Other with the given parameters.

pub fn is_unique_key_error<View, C>(&self, connection: &C) -> boolwhere
View: View,
C: HasSchema,

Returns true if this error is a Error::UniqueKeyViolation from View.

pub fn conflicting_document<Collection>(&self) -> Option<Header>where
Collection: Collection,

Returns the header of the conflicting document if this error is a Error::DocumentConflict from Collection.