Module bonsaidb::core::key::time

Key implementations for time types.

Modules

limited

Types for storing limited-precision Durations.

Structs

DeltaNotRepresentable

An error that indicates that the stored timestamp is unable to be converted to the destination type without losing data.

Enums

TimeError

Errors that can arise from parsing times serialized with Key.

Type Definitions

Days

A signed duration of time represented in days (86,400 seconds). Internally, the number of days is represented as an i32, allowing a range of +/- ~5.88 million years.

DaysSinceUnixEpoch

A timestamp stored as the number of days (86,400 seconds) relative to UnixEpoch. Internally, the number of days is represented as an i32, allowing a range of +/- ~5.88 million years.

Hours

A signed duration of time represented in hours (3,600 seconds). Internally, the number of hours is represented as an i32, allowing a range of +/- ~245,147 years.

HoursSinceUnixEpoch

A timestamp stored as the number of hours (3,600 seconds) relative to UnixEpoch. Internally, the number of hours is represented as an i32, allowing a range of +/- ~245,147 years.

Microseconds

A signed duration of time represented in microseconds (1/1,000,000th of a second). Internally, the number of microseconds is represented as an i64, allowing a range of +/- ~292,471 years.

MicrosecondsSinceUnixEpoch

A timestamp stored as the number of microseconds (1/1,000,000th of a second) relative to UnixEpoch. Internally, the number of microseconds is represented as an i64, allowing a range of +/- ~292,471 years.

Milliseconds

A signed duration of time represented in milliseconds (1/1,000th of a second). Internally, the number of milliseconds is represented as an i64, allowing a range of +/- ~292.5 million years.

MillisecondsSinceUnixEpoch

A timestamp stored as the number of milliseconds (1/1,000th of a second) relative to UnixEpoch. Internally, the number of milliseconds is represented as an i64, allowing a range of +/- ~292.5 million years.

Minutes

A signed duration of time represented in minutes (60 seconds). Internally, the number of minutes is represented as an i32, allowing a range of +/- ~4,086 years.

MinutesSinceUnixEpoch

A timestamp stored as the number of minutes (60 seconds) relative to UnixEpoch. Internally, the number of minutes is represented as an i32, allowing a range of +/- ~4,086 years.

Nanoseconds

A signed duration of time represented in nanoseconds (1/1,000,000,000th of a second). Internally, the number of microseconds is represented as an i64, allowing a range of +/- ~292.5 years.

NanosecondsSinceUnixEpoch

A timestamp stored as the number of nanoseconds (1/1,000,000,000th of a second) relative to UnixEpoch. Internally, the number of microseconds is represented as an i64, allowing a range of +/- ~292.5 years.

Seconds

A signed duration of time represented in seconds (with no partial subseconds). Internally, the number of seconds is represented as an i64, allowing a range of +/- ~21 times the age of the universe.

SecondsSinceUnixEpoch

A timestamp stored as the number of seconds (with no partial subseconds) relative to UnixEpoch. Internally, the number of seconds is represented as an i64, allowing a range of +/- ~21 times the age of the universe.

TimestampAsDays

A timestamp stored as the number of days (86,400 seconds) relative to BonsaiEpoch. Internally, the number of days is represented as an i32, allowing a range of +/- ~5.88 million years.

TimestampAsHours

A timestamp stored as the number of hours (3,600 seconds) relative to BonsaiEpoch. Internally, the number of hours is represented as an i32, allowing a range of +/- ~245,147 years.

TimestampAsMicroseconds

A timestamp stored as the number of microseconds (1/1,000,000th of a second) relative to BonsaiEpoch. Internally, the number of microseconds is represented as an i64, allowing a range of +/- ~292,471 years.

TimestampAsMilliseconds

A timestamp stored as the number of milliseconds (1/1,000th of a second) relative to BonsaiEpoch. Internally, the number of milliseconds is represented as an i64, allowing a range of +/- ~292.5 million years.

TimestampAsMinutes

A timestamp stored as the number of minutes (60 seconds) relative to BonsaiEpoch. Internally, the number of minutes is represented as an i32, allowing a range of +/- ~4,086 years.

TimestampAsNanoseconds

A timestamp stored as the number of nanoseconds (1/1,000,000,000th of a second) relative to BonsaiEpoch. Internally, the number of microseconds is represented as an i64, allowing a range of +/- ~292.5 years.

TimestampAsSeconds

A timestamp stored as the number of seconds (with no partial subseconds) relative to BonsaiEpoch. Internally, the number of seconds is represented as an i64, allowing a range of +/- ~21 times the age of the universe.

TimestampAsWeeks

A timestamp stored as the number of weeks (604,800 seconds) relative to BonsaiEpoch. Internally, the number of weeks is represented as an i32, allowing a range of +/- ~41.18 million years.

Weeks

A signed duration of time represented in weeks (604,800 seconds). Internally, the number of weeks is represented as an i32, allowing a range of +/- ~41.18 million years.

WeeksSinceUnixEpoch

A timestamp stored as the number of weeks (604,800 seconds) relative to UnixEpoch. Internally, the number of weeks is represented as an i32, allowing a range of +/- ~41.18 million years.