Enum bonsaidb::core::networking::DatabaseRequest

pub enum DatabaseRequest {

    Get {
        collection: CollectionName,
        id: u64,
    },
    GetMultiple {
        collection: CollectionName,
        ids: Vec<u64, Global>,
    },
    List {
        collection: CollectionName,
        ids: Range<u64>,
        order: Sort,
        limit: Option<usize>,
    },
    Query {
        view: ViewName,
        key: Option<QueryKey<Vec<u8, Global>>>,
        order: Sort,
        limit: Option<usize>,
        access_policy: AccessPolicy,
        with_docs: bool,
    },
    Reduce {
        view: ViewName,
        key: Option<QueryKey<Vec<u8, Global>>>,
        access_policy: AccessPolicy,
        grouped: bool,
    },
    DeleteDocs {
        view: ViewName,
        key: Option<QueryKey<Vec<u8, Global>>>,
        access_policy: AccessPolicy,
    },
    ApplyTransaction {
        transaction: Transaction,
    },
    ListExecutedTransactions {
        starting_id: Option<u64>,
        result_limit: Option<usize>,
    },
    LastTransactionId,
    CreateSubscriber,
    Publish {
        topic: String,
        payload: Vec<u8, Global>,
    },
    PublishToAll {
        topics: Vec<String, Global>,
        payload: Vec<u8, Global>,
    },
    SubscribeTo {
        subscriber_id: u64,
        topic: String,
    },
    UnsubscribeFrom {
        subscriber_id: u64,
        topic: String,
    },
    UnregisterSubscriber {
        subscriber_id: u64,
    },
    ExecuteKeyOperation(KeyOperation),
    CompactCollection {
        name: CollectionName,
    },
    CompactKeyValueStore,
    Compact,

}

A database-related request.

Variants

Get

Retrieve a single document.

Fields of Get

collection: CollectionName

The collection of the document.

id: u64

The id of the document.

GetMultiple

Retrieve multiple documents.

Fields of GetMultiple

collection: CollectionName

The collection of the documents.

ids: Vec<u64, Global>

The ids of the documents.

List

Retrieve multiple documents.

Fields of List

collection: CollectionName

The collection of the documents.

ids: Range<u64>

The range of ids to list.

order: Sort

The order for the query into the collection.

limit: Option<usize>

The maximum number of results to return.

Query

Queries a view.

Fields of Query

view: ViewName

The name of the view.

key: Option<QueryKey<Vec<u8, Global>>>

The filter for the view.

order: Sort

The order for the query into the view.

limit: Option<usize>

The maximum number of results to return.

access_policy: AccessPolicy

The access policy for the query.

with_docs: bool

If true, DatabaseResponse::ViewMappingsWithDocs will be returned. If false, DatabaseResponse::ViewMappings will be returned.

Reduce

Reduces a view.

Fields of Reduce

view: ViewName

The name of the view.

key: Option<QueryKey<Vec<u8, Global>>>

The filter for the view.

access_policy: AccessPolicy

The access policy for the query.

grouped: bool

Whether to return a single value or values grouped by unique key. If true, DatabaseResponse::ViewGroupedReduction will be returned. If false, DatabaseResponse::ViewReduction is returned.

DeleteDocs

Deletes the associated documents resulting from the view query.

Fields of DeleteDocs

view: ViewName

The name of the view.

key: Option<QueryKey<Vec<u8, Global>>>

The filter for the view.

access_policy: AccessPolicy

The access policy for the query.

ApplyTransaction

Applies a transaction.

Fields of ApplyTransaction

transaction: Transaction

The trasnaction to apply.

ListExecutedTransactions

Lists executed transactions.

Fields of ListExecutedTransactions

starting_id: Option<u64>

The starting transaction id.

result_limit: Option<usize>

The maximum number of results.

LastTransactionId

Queries the last transaction id.

CreateSubscriber

Creates a PubSub Subscriber

Publish

Publishes payload to all subscribers of topic.

Fields of Publish

topic: String

The topics to publish to.

payload: Vec<u8, Global>

The payload to publish.

PublishToAll

Publishes payload to all subscribers of all topics.

Fields of PublishToAll

topics: Vec<String, Global>

The topics to publish to.

payload: Vec<u8, Global>

The payload to publish.

SubscribeTo

Subscribes subscriber_id to messages for topic.

Fields of SubscribeTo

subscriber_id: u64

The id of the Subscriber.

topic: String

The topic to subscribe to.

UnsubscribeFrom

Unsubscribes subscriber_id from messages for topic.

Fields of UnsubscribeFrom

subscriber_id: u64

The id of the Subscriber.

topic: String

The topic to unsubscribe from.

UnregisterSubscriber

Unregisters the subscriber.

Fields of UnregisterSubscriber

subscriber_id: u64

The id of the Subscriber.

ExecuteKeyOperation(KeyOperation)

Excutes a key-value store operation.

Tuple Fields of ExecuteKeyOperation

0: KeyOperation

CompactCollection

Compacts the collection.

Fields of CompactCollection

name: CollectionName

The name of the collection to compact.

CompactKeyValueStore

Compacts the key-value store.

Compact

Compacts the entire database.