Module bonsaidb::core::pubsub

Types for Publish/Subscribe (PubSub) messaging.

Macros

define_async_pubsub_test_suite

Expands into a suite of pubsub unit tests using the passed type as the test harness.

define_blocking_pubsub_test_suite

Expands into a suite of pubsub unit tests using the passed type as the test harness.

Structs

Disconnected

The Receiver was disconnected

Receiver

Receiver of PubSub Messages.

Enums

TryReceiveError

An error occurred trying to receive a message.

Traits

AsyncPubSub

Publishes and Subscribes to messages on topics.

AsyncSubscriber

A subscriber to one or more topics.

PubSub

Publishes and Subscribes to messages on topics.

Subscriber

A subscriber to one or more topics.