Enum bonsaidb::core::schema::NamedReference

pub enum NamedReference<'a, Id> {
    Name(Cow<'a, str>),
    Id(DocumentId),
    Key(Id),
}

A reference to a collection that has a unique name view.

Variants

Name(Cow<'a, str>)

An entity’s name.

Id(DocumentId)

A document id.

Key(Id)

A document id.

Implementations

impl<'a, Id> NamedReference<'a, Id>where
Id: for<'k> Key<'k>,

pub fn into_owned(self) -> NamedReference<'static, Id>

Converts this reference to an owned reference with a 'static lifetime.

pub fn id<Col, Cn>(
&self,
connection: &Cn
) -> Result<Option<<Col as Collection>::PrimaryKey>, Error>where
Col: NamedCollection<PrimaryKey = Id>,
Cn: Connection,

Returns this reference’s id. If the reference is a name, the NamedCollection::ByNameView is queried for the id.

pub async fn id_async<Col, Cn>(
&self,
connection: &Cn
) -> impl Future<Output = Result<Option<<Col as Collection>::PrimaryKey>, Error>>where
Col: NamedCollection<PrimaryKey = Id>,
Cn: AsyncConnection,

Returns this reference’s id. If the reference is a name, the NamedCollection::ByNameView is queried for the id.