Trait bonsaidb::core::schema::Qualified

pub trait Qualified: Display {
    fn new<A, N>(authority: A, name: N) -> Self
    where
        A: Into<Authority>,
        N: Into<Name>;

    fn private<N>(name: N) -> Self
    where
        N: Into<Name>,
    { ... }
    fn parse_encoded(schema_name: &str) -> Result<Self, InvalidNameError> { ... }
    fn encoded(&self) -> String { ... }
}

Functions for creating qualified names

Required Methods

fn new<A, N>(authority: A, name: N) -> Self where
A: Into<Authority>,
N: Into<Name>,

Creates a new qualified name.

Provided Methods

fn private<N>(name: N) -> Self where
N: Into<Name>,

Creates a name that is not meant to be shared with other developers or projects.

fn parse_encoded(schema_name: &str) -> Result<Self, InvalidNameError>

Parses a schema name that was previously encoded via Self::encoded().

Errors

Returns InvalidNameError if the name contains invalid escape sequences or contains more than two periods.

fn encoded(&self) -> String

Encodes this schema name such that the authority and name can be safely parsed using Self::parse_encoded.