Module bonsaidb::core::schema::view

Types for defining map/reduce-powered Views.

Modules

map

Types for defining a Map within a View.

Structs

CompositeKeyError

An error occurred inside of one of the composite key fields.

IncorrectByteLength

An error that indicates an unexpected number of bytes were present.

UnknownEnumVariant

An error that indicates an unexpected enum variant value was found.

Enums

Error

Errors that arise when interacting with views.

Traits

CollectionView

A View for a Collection that stores Serde-compatible documents. The only difference between implmementing this and View is that the map function receives a CollectionDocument instead of a Document.

DefaultViewSerialization

A default serialization strategy for views. Uses equivalent settings as DefaultSerialization.

EnumKey

Adds Key support to an enum. Requires implementing ToPrimitive and FromPrimitive, or using a crate like num-derive to do it automatically. Take care when using enums as keys: if the order changes or if the meaning of existing numerical values changes, make sure to update any related views’ version number to ensure the values are re-evaluated.

Key

A trait that enables a type to convert itself to a big-endian/network byte order.

Serialized

Wraps a View with serialization to erase the associated types

SerializedView

A View with additional tyes and logic to handle serializing view values.

View

A map/reduce powered indexing and aggregation schema.

Type Definitions

MapResult

A type alias for the result of View::map().