Enum bonsaidb::AnyDatabase

pub enum AnyDatabase<B: Backend = NoBackend> {
    Local(ServerDatabase<B>),
    Networked(RemoteDatabase<B::CustomApi>),
}

A database connection that can be either from a local server or a server over a network connection.

Variants

Local(ServerDatabase<B>)

Tuple Fields

0: ServerDatabase<B>

A local database.

Networked(RemoteDatabase<B::CustomApi>)

Tuple Fields

0: RemoteDatabase<B::CustomApi>

A networked database accessed with a Client.