Struct bonsaidb::files::direct::BufferedAppend

pub struct BufferedAppend<'a, Config, Database>where
    Config: FileConfig,
    Database: Connection + Clone,{ /* private fields */ }

A buffered std::io::Write and std::io::Seek implementor for a File.

Implementations

impl<'a, Config, Database> BufferedAppend<'a, Config, Database>where
Config: FileConfig,
Database: Connection + Clone,

pub fn set_buffer_size(&mut self, capacity: usize) -> Result<(), Error>

Sets the size of the buffer. For optimal use, this should be a multiple of Config::BLOCK_SIZE.

If any data is already buffered, it will be flushed before the buffer is resized.