Module bonsaidb::keystorage::s3::aws_sdk_s3::client::fluent_builders

Utilities to ergonomically construct a request to the service.

Fluent builders are created through the Client by calling one if its operation methods. After parameters are set using the builder methods, the send method can be called to initiate the request.

Structs

AbortMultipartUpload

Fluent builder constructing a request to AbortMultipartUpload.

CompleteMultipartUpload

Fluent builder constructing a request to CompleteMultipartUpload.

CopyObject

Fluent builder constructing a request to CopyObject.

CreateBucket

Fluent builder constructing a request to CreateBucket.

CreateMultipartUpload

Fluent builder constructing a request to CreateMultipartUpload.

DeleteBucket

Fluent builder constructing a request to DeleteBucket.

DeleteBucketAnalyticsConfiguration

Fluent builder constructing a request to DeleteBucketAnalyticsConfiguration.

DeleteBucketCors

Fluent builder constructing a request to DeleteBucketCors.

DeleteBucketEncryption

Fluent builder constructing a request to DeleteBucketEncryption.

DeleteBucketIntelligentTieringConfiguration

Fluent builder constructing a request to DeleteBucketIntelligentTieringConfiguration.

DeleteBucketInventoryConfiguration

Fluent builder constructing a request to DeleteBucketInventoryConfiguration.

DeleteBucketLifecycle

Fluent builder constructing a request to DeleteBucketLifecycle.

DeleteBucketMetricsConfiguration

Fluent builder constructing a request to DeleteBucketMetricsConfiguration.

DeleteBucketOwnershipControls

Fluent builder constructing a request to DeleteBucketOwnershipControls.

DeleteBucketPolicy

Fluent builder constructing a request to DeleteBucketPolicy.

DeleteBucketReplication

Fluent builder constructing a request to DeleteBucketReplication.

DeleteBucketTagging

Fluent builder constructing a request to DeleteBucketTagging.

DeleteBucketWebsite

Fluent builder constructing a request to DeleteBucketWebsite.

DeleteObject

Fluent builder constructing a request to DeleteObject.

DeleteObjectTagging

Fluent builder constructing a request to DeleteObjectTagging.

DeleteObjects

Fluent builder constructing a request to DeleteObjects.

DeletePublicAccessBlock

Fluent builder constructing a request to DeletePublicAccessBlock.

GetBucketAccelerateConfiguration

Fluent builder constructing a request to GetBucketAccelerateConfiguration.

GetBucketAcl

Fluent builder constructing a request to GetBucketAcl.

GetBucketAnalyticsConfiguration

Fluent builder constructing a request to GetBucketAnalyticsConfiguration.

GetBucketCors

Fluent builder constructing a request to GetBucketCors.

GetBucketEncryption

Fluent builder constructing a request to GetBucketEncryption.

GetBucketIntelligentTieringConfiguration

Fluent builder constructing a request to GetBucketIntelligentTieringConfiguration.

GetBucketInventoryConfiguration

Fluent builder constructing a request to GetBucketInventoryConfiguration.

GetBucketLifecycleConfiguration

Fluent builder constructing a request to GetBucketLifecycleConfiguration.

GetBucketLocation

Fluent builder constructing a request to GetBucketLocation.

GetBucketLogging

Fluent builder constructing a request to GetBucketLogging.

GetBucketMetricsConfiguration

Fluent builder constructing a request to GetBucketMetricsConfiguration.

GetBucketNotificationConfiguration

Fluent builder constructing a request to GetBucketNotificationConfiguration.

GetBucketOwnershipControls

Fluent builder constructing a request to GetBucketOwnershipControls.

GetBucketPolicy

Fluent builder constructing a request to GetBucketPolicy.

GetBucketPolicyStatus

Fluent builder constructing a request to GetBucketPolicyStatus.

GetBucketReplication

Fluent builder constructing a request to GetBucketReplication.

GetBucketRequestPayment

Fluent builder constructing a request to GetBucketRequestPayment.

GetBucketTagging

Fluent builder constructing a request to GetBucketTagging.

GetBucketVersioning

Fluent builder constructing a request to GetBucketVersioning.

GetBucketWebsite

Fluent builder constructing a request to GetBucketWebsite.

GetObject

Fluent builder constructing a request to GetObject.

GetObjectAcl

Fluent builder constructing a request to GetObjectAcl.

GetObjectLegalHold

Fluent builder constructing a request to GetObjectLegalHold.

GetObjectLockConfiguration

Fluent builder constructing a request to GetObjectLockConfiguration.

GetObjectRetention

Fluent builder constructing a request to GetObjectRetention.

GetObjectTagging

Fluent builder constructing a request to GetObjectTagging.

GetObjectTorrent

Fluent builder constructing a request to GetObjectTorrent.

GetPublicAccessBlock

Fluent builder constructing a request to GetPublicAccessBlock.

HeadBucket

Fluent builder constructing a request to HeadBucket.

HeadObject

Fluent builder constructing a request to HeadObject.

ListBucketAnalyticsConfigurations

Fluent builder constructing a request to ListBucketAnalyticsConfigurations.

ListBucketIntelligentTieringConfigurations

Fluent builder constructing a request to ListBucketIntelligentTieringConfigurations.

ListBucketInventoryConfigurations

Fluent builder constructing a request to ListBucketInventoryConfigurations.

ListBucketMetricsConfigurations

Fluent builder constructing a request to ListBucketMetricsConfigurations.

ListBuckets

Fluent builder constructing a request to ListBuckets.

ListMultipartUploads

Fluent builder constructing a request to ListMultipartUploads.

ListObjectVersions

Fluent builder constructing a request to ListObjectVersions.

ListObjects

Fluent builder constructing a request to ListObjects.

ListObjectsV2

Fluent builder constructing a request to ListObjectsV2.

ListParts

Fluent builder constructing a request to ListParts.

PutBucketAccelerateConfiguration

Fluent builder constructing a request to PutBucketAccelerateConfiguration.

PutBucketAcl

Fluent builder constructing a request to PutBucketAcl.

PutBucketAnalyticsConfiguration

Fluent builder constructing a request to PutBucketAnalyticsConfiguration.

PutBucketCors

Fluent builder constructing a request to PutBucketCors.

PutBucketEncryption

Fluent builder constructing a request to PutBucketEncryption.

PutBucketIntelligentTieringConfiguration

Fluent builder constructing a request to PutBucketIntelligentTieringConfiguration.

PutBucketInventoryConfiguration

Fluent builder constructing a request to PutBucketInventoryConfiguration.

PutBucketLifecycleConfiguration

Fluent builder constructing a request to PutBucketLifecycleConfiguration.

PutBucketLogging

Fluent builder constructing a request to PutBucketLogging.

PutBucketMetricsConfiguration

Fluent builder constructing a request to PutBucketMetricsConfiguration.

PutBucketNotificationConfiguration

Fluent builder constructing a request to PutBucketNotificationConfiguration.

PutBucketOwnershipControls

Fluent builder constructing a request to PutBucketOwnershipControls.

PutBucketPolicy

Fluent builder constructing a request to PutBucketPolicy.

PutBucketReplication

Fluent builder constructing a request to PutBucketReplication.

PutBucketRequestPayment

Fluent builder constructing a request to PutBucketRequestPayment.

PutBucketTagging

Fluent builder constructing a request to PutBucketTagging.

PutBucketVersioning

Fluent builder constructing a request to PutBucketVersioning.

PutBucketWebsite

Fluent builder constructing a request to PutBucketWebsite.

PutObject

Fluent builder constructing a request to PutObject.

PutObjectAcl

Fluent builder constructing a request to PutObjectAcl.

PutObjectLegalHold

Fluent builder constructing a request to PutObjectLegalHold.

PutObjectLockConfiguration

Fluent builder constructing a request to PutObjectLockConfiguration.

PutObjectRetention

Fluent builder constructing a request to PutObjectRetention.

PutObjectTagging

Fluent builder constructing a request to PutObjectTagging.

PutPublicAccessBlock

Fluent builder constructing a request to PutPublicAccessBlock.

RestoreObject

Fluent builder constructing a request to RestoreObject.

SelectObjectContent

Fluent builder constructing a request to SelectObjectContent.

UploadPart

Fluent builder constructing a request to UploadPart.

UploadPartCopy

Fluent builder constructing a request to UploadPartCopy.

WriteGetObjectResponse

Fluent builder constructing a request to WriteGetObjectResponse.