Struct bonsaidb::keystorage::s3::aws_sdk_s3::config::Builder

pub struct Builder { /* private fields */ }

Builder for creating a Config.

Implementations

impl Builder

pub fn new() -> Builder

Constructs a config builder.

pub fn app_name(self, app_name: AppName) -> Builder

Sets the name of the app that is using the client.

This optional name is used to identify the application in the user agent that gets sent along with requests.

pub fn set_app_name(&mut self, app_name: Option<AppName>) -> &mut Builder

Sets the name of the app that is using the client.

This optional name is used to identify the application in the user agent that gets sent along with requests.

pub fn timeout_config(self, timeout_config: Config) -> Builder

Set the timeout_config for the builder

Examples

use aws_sdk_s3::config::Config;
use aws_smithy_types::{timeout, tristate::TriState};

let api_timeouts = timeout::Api::new()
    .with_call_attempt_timeout(TriState::Set(Duration::from_secs(1)));
let timeout_config = timeout::Config::new()
    .with_api_timeouts(api_timeouts);
let config = Config::builder().timeout_config(timeout_config).build();

pub fn set_timeout_config(
&mut self,
timeout_config: Option<Config>
) -> &mut Builder

Set the timeout_config for the builder

Examples

use aws_sdk_s3::config::{Builder, Config};
use aws_smithy_types::{timeout, tristate::TriState};

fn set_request_timeout(builder: &mut Builder) {
    let api_timeouts = timeout::Api::new()
        .with_call_attempt_timeout(TriState::Set(Duration::from_secs(1)));
    let timeout_config = timeout::Config::new()
        .with_api_timeouts(api_timeouts);
    builder.set_timeout_config(Some(timeout_config));
}

let mut builder = Config::builder();
set_request_timeout(&mut builder);
let config = builder.build();

pub fn sleep_impl(self, sleep_impl: Arc<dyn AsyncSleep + 'static>) -> Builder

Set the sleep_impl for the builder

Examples

use aws_sdk_s3::config::Config;
use aws_smithy_async::rt::sleep::AsyncSleep;
use aws_smithy_async::rt::sleep::Sleep;

#[derive(Debug)]
pub struct ForeverSleep;

impl AsyncSleep for ForeverSleep {
    fn sleep(&self, duration: std::time::Duration) -> Sleep {
        Sleep::new(std::future::pending())
    }
}

let sleep_impl = std::sync::Arc::new(ForeverSleep);
let config = Config::builder().sleep_impl(sleep_impl).build();

pub fn set_sleep_impl(
&mut self,
sleep_impl: Option<Arc<dyn AsyncSleep + 'static>>
) -> &mut Builder

Set the sleep_impl for the builder

Examples

use aws_sdk_s3::config::{Builder, Config};
use aws_smithy_async::rt::sleep::AsyncSleep;
use aws_smithy_async::rt::sleep::Sleep;

#[derive(Debug)]
pub struct ForeverSleep;

impl AsyncSleep for ForeverSleep {
    fn sleep(&self, duration: std::time::Duration) -> Sleep {
        Sleep::new(std::future::pending())
    }
}

fn set_never_ending_sleep_impl(builder: &mut Builder) {
    let sleep_impl = std::sync::Arc::new(ForeverSleep);
    builder.set_sleep_impl(Some(sleep_impl));
}

let mut builder = Config::builder();
set_never_ending_sleep_impl(&mut builder);
let config = builder.build();

pub fn retry_config(self, retry_config: RetryConfig) -> Builder

Set the retry_config for the builder

Examples

use aws_sdk_s3::config::Config;
use aws_smithy_types::retry::RetryConfig;

let retry_config = RetryConfig::new().with_max_attempts(5);
let config = Config::builder().retry_config(retry_config).build();

pub fn set_retry_config(
&mut self,
retry_config: Option<RetryConfig>
) -> &mut Builder

Set the retry_config for the builder

Examples

use aws_sdk_s3::config::{Builder, Config};
use aws_smithy_types::retry::RetryConfig;

fn disable_retries(builder: &mut Builder) {
    let retry_config = RetryConfig::new().with_max_attempts(1);
    builder.set_retry_config(Some(retry_config));
}

let mut builder = Config::builder();
disable_retries(&mut builder);
let config = builder.build();

pub fn endpoint_resolver(
self,
endpoint_resolver: impl ResolveAwsEndpoint + 'static
) -> Builder

Sets the endpoint resolver to use when making requests.

pub fn region(self, region: impl Into<Option<Region>>) -> Builder

Sets the AWS region to use when making requests.

Examples

use aws_types::region::Region;
use aws_sdk_s3::config::{Builder, Config};

let config = aws_sdk_s3::Config::builder()
    .region(Region::new("us-east-1"))
    .build();

pub fn credentials_provider(
self,
credentials_provider: impl ProvideCredentials + 'static
) -> Builder

Sets the credentials provider for this service

pub fn set_credentials_provider(
&mut self,
credentials_provider: Option<SharedCredentialsProvider>
) -> &mut Builder

Sets the credentials provider for this service

pub fn build(self) -> Config

Builds a Config.