Struct bonsaidb::keystorage::s3::aws_sdk_s3::input::write_get_object_response_input::Builder

#[non_exhaustive]
pub struct Builder { /* private fields */ }

A builder for WriteGetObjectResponseInput

Implementations

impl Builder

pub fn request_route(self, input: impl Into<String>) -> Builder

Route prefix to the HTTP URL generated.

pub fn set_request_route(self, input: Option<String>) -> Builder

Route prefix to the HTTP URL generated.

pub fn request_token(self, input: impl Into<String>) -> Builder

A single use encrypted token that maps WriteGetObjectResponse to the end user GetObject request.

pub fn set_request_token(self, input: Option<String>) -> Builder

A single use encrypted token that maps WriteGetObjectResponse to the end user GetObject request.

pub fn body(self, input: ByteStream) -> Builder

The object data.

pub fn set_body(self, input: Option<ByteStream>) -> Builder

The object data.

pub fn status_code(self, input: i32) -> Builder

The integer status code for an HTTP response of a corresponding GetObject request.

Status Codes

200 - OK

206 - Partial Content

304 - Not Modified

400 - Bad Request

401 - Unauthorized

403 - Forbidden

404 - Not Found

405 - Method Not Allowed

409 - Conflict

411 - Length Required

412 - Precondition Failed

416 - Range Not Satisfiable

500 - Internal Server Error

503 - Service Unavailable

pub fn set_status_code(self, input: Option<i32>) -> Builder

The integer status code for an HTTP response of a corresponding GetObject request.

Status Codes

200 - OK

206 - Partial Content

304 - Not Modified

400 - Bad Request

401 - Unauthorized

403 - Forbidden

404 - Not Found

405 - Method Not Allowed

409 - Conflict

411 - Length Required

412 - Precondition Failed

416 - Range Not Satisfiable

500 - Internal Server Error

503 - Service Unavailable

pub fn error_code(self, input: impl Into<String>) -> Builder

A string that uniquely identifies an error condition. Returned in the tag of the error XML response for a corresponding GetObject call. Cannot be used with a successful StatusCode header or when the transformed object is provided in the body. All error codes from S3 are sentence-cased. Regex value is "^[A-Z][a-zA-Z]+$".

pub fn set_error_code(self, input: Option<String>) -> Builder

A string that uniquely identifies an error condition. Returned in the tag of the error XML response for a corresponding GetObject call. Cannot be used with a successful StatusCode header or when the transformed object is provided in the body. All error codes from S3 are sentence-cased. Regex value is "^[A-Z][a-zA-Z]+$".

pub fn error_message(self, input: impl Into<String>) -> Builder

Contains a generic description of the error condition. Returned in the tag of the error XML response for a corresponding GetObject call. Cannot be used with a successful StatusCode header or when the transformed object is provided in body.

pub fn set_error_message(self, input: Option<String>) -> Builder

Contains a generic description of the error condition. Returned in the tag of the error XML response for a corresponding GetObject call. Cannot be used with a successful StatusCode header or when the transformed object is provided in body.

pub fn accept_ranges(self, input: impl Into<String>) -> Builder

Indicates that a range of bytes was specified.

pub fn set_accept_ranges(self, input: Option<String>) -> Builder

Indicates that a range of bytes was specified.

pub fn cache_control(self, input: impl Into<String>) -> Builder

Specifies caching behavior along the request/reply chain.

pub fn set_cache_control(self, input: Option<String>) -> Builder

Specifies caching behavior along the request/reply chain.

pub fn content_disposition(self, input: impl Into<String>) -> Builder

Specifies presentational information for the object.

pub fn set_content_disposition(self, input: Option<String>) -> Builder

Specifies presentational information for the object.

pub fn content_encoding(self, input: impl Into<String>) -> Builder

Specifies what content encodings have been applied to the object and thus what decoding mechanisms must be applied to obtain the media-type referenced by the Content-Type header field.

pub fn set_content_encoding(self, input: Option<String>) -> Builder

Specifies what content encodings have been applied to the object and thus what decoding mechanisms must be applied to obtain the media-type referenced by the Content-Type header field.

pub fn content_language(self, input: impl Into<String>) -> Builder

The language the content is in.

pub fn set_content_language(self, input: Option<String>) -> Builder

The language the content is in.

pub fn content_length(self, input: i64) -> Builder

The size of the content body in bytes.

pub fn set_content_length(self, input: Option<i64>) -> Builder

The size of the content body in bytes.

pub fn content_range(self, input: impl Into<String>) -> Builder

The portion of the object returned in the response.

pub fn set_content_range(self, input: Option<String>) -> Builder

The portion of the object returned in the response.

pub fn content_type(self, input: impl Into<String>) -> Builder

A standard MIME type describing the format of the object data.

pub fn set_content_type(self, input: Option<String>) -> Builder

A standard MIME type describing the format of the object data.

pub fn delete_marker(self, input: bool) -> Builder

Specifies whether an object stored in Amazon S3 is (true) or is not (false) a delete marker.

pub fn set_delete_marker(self, input: Option<bool>) -> Builder

Specifies whether an object stored in Amazon S3 is (true) or is not (false) a delete marker.

pub fn e_tag(self, input: impl Into<String>) -> Builder

An opaque identifier assigned by a web server to a specific version of a resource found at a URL.

pub fn set_e_tag(self, input: Option<String>) -> Builder

An opaque identifier assigned by a web server to a specific version of a resource found at a URL.

pub fn expires(self, input: DateTime) -> Builder

The date and time at which the object is no longer cacheable.

pub fn set_expires(self, input: Option<DateTime>) -> Builder

The date and time at which the object is no longer cacheable.

pub fn expiration(self, input: impl Into<String>) -> Builder

If object stored in Amazon S3 expiration is configured (see PUT Bucket lifecycle) it includes expiry-date and rule-id key-value pairs providing object expiration information. The value of the rule-id is URL encoded.

pub fn set_expiration(self, input: Option<String>) -> Builder

If object stored in Amazon S3 expiration is configured (see PUT Bucket lifecycle) it includes expiry-date and rule-id key-value pairs providing object expiration information. The value of the rule-id is URL encoded.

pub fn last_modified(self, input: DateTime) -> Builder

The date and time that the object was last modified.

pub fn set_last_modified(self, input: Option<DateTime>) -> Builder

The date and time that the object was last modified.

pub fn missing_meta(self, input: i32) -> Builder

Set to the number of metadata entries not returned in x-amz-meta headers. This can happen if you create metadata using an API like SOAP that supports more flexible metadata than the REST API. For example, using SOAP, you can create metadata whose values are not legal HTTP headers.

pub fn set_missing_meta(self, input: Option<i32>) -> Builder

Set to the number of metadata entries not returned in x-amz-meta headers. This can happen if you create metadata using an API like SOAP that supports more flexible metadata than the REST API. For example, using SOAP, you can create metadata whose values are not legal HTTP headers.

pub fn metadata(self, k: impl Into<String>, v: impl Into<String>) -> Builder

Adds a key-value pair to metadata.

To override the contents of this collection use set_metadata.

A map of metadata to store with the object in S3.

pub fn set_metadata(
self,
input: Option<HashMap<String, String, RandomState>>
) -> Builder

A map of metadata to store with the object in S3.

pub fn object_lock_mode(self, input: ObjectLockMode) -> Builder

Indicates whether an object stored in Amazon S3 has Object Lock enabled. For more information about S3 Object Lock, see Object Lock.

pub fn set_object_lock_mode(self, input: Option<ObjectLockMode>) -> Builder

Indicates whether an object stored in Amazon S3 has Object Lock enabled. For more information about S3 Object Lock, see Object Lock.

pub fn object_lock_legal_hold_status(
self,
input: ObjectLockLegalHoldStatus
) -> Builder

Indicates whether an object stored in Amazon S3 has an active legal hold.

pub fn set_object_lock_legal_hold_status(
self,
input: Option<ObjectLockLegalHoldStatus>
) -> Builder

Indicates whether an object stored in Amazon S3 has an active legal hold.

pub fn object_lock_retain_until_date(self, input: DateTime) -> Builder

The date and time when Object Lock is configured to expire.

pub fn set_object_lock_retain_until_date(
self,
input: Option<DateTime>
) -> Builder

The date and time when Object Lock is configured to expire.

pub fn parts_count(self, input: i32) -> Builder

The count of parts this object has.

pub fn set_parts_count(self, input: Option<i32>) -> Builder

The count of parts this object has.

pub fn replication_status(self, input: ReplicationStatus) -> Builder

Indicates if request involves bucket that is either a source or destination in a Replication rule. For more information about S3 Replication, see Replication.

pub fn set_replication_status(self, input: Option<ReplicationStatus>) -> Builder

Indicates if request involves bucket that is either a source or destination in a Replication rule. For more information about S3 Replication, see Replication.

pub fn request_charged(self, input: RequestCharged) -> Builder

If present, indicates that the requester was successfully charged for the request.

pub fn set_request_charged(self, input: Option<RequestCharged>) -> Builder

If present, indicates that the requester was successfully charged for the request.

pub fn restore(self, input: impl Into<String>) -> Builder

Provides information about object restoration operation and expiration time of the restored object copy.

pub fn set_restore(self, input: Option<String>) -> Builder

Provides information about object restoration operation and expiration time of the restored object copy.

pub fn server_side_encryption(self, input: ServerSideEncryption) -> Builder

The server-side encryption algorithm used when storing requested object in Amazon S3 (for example, AES256, aws:kms).

pub fn set_server_side_encryption(
self,
input: Option<ServerSideEncryption>
) -> Builder

The server-side encryption algorithm used when storing requested object in Amazon S3 (for example, AES256, aws:kms).

pub fn sse_customer_algorithm(self, input: impl Into<String>) -> Builder

Encryption algorithm used if server-side encryption with a customer-provided encryption key was specified for object stored in Amazon S3.

pub fn set_sse_customer_algorithm(self, input: Option<String>) -> Builder

Encryption algorithm used if server-side encryption with a customer-provided encryption key was specified for object stored in Amazon S3.

pub fn ssekms_key_id(self, input: impl Into<String>) -> Builder

If present, specifies the ID of the Amazon Web Services Key Management Service (Amazon Web Services KMS) symmetric customer managed key that was used for stored in Amazon S3 object.

pub fn set_ssekms_key_id(self, input: Option<String>) -> Builder

If present, specifies the ID of the Amazon Web Services Key Management Service (Amazon Web Services KMS) symmetric customer managed key that was used for stored in Amazon S3 object.

pub fn sse_customer_key_md5(self, input: impl Into<String>) -> Builder

128-bit MD5 digest of customer-provided encryption key used in Amazon S3 to encrypt data stored in S3. For more information, see Protecting data using server-side encryption with customer-provided encryption keys (SSE-C).

pub fn set_sse_customer_key_md5(self, input: Option<String>) -> Builder

128-bit MD5 digest of customer-provided encryption key used in Amazon S3 to encrypt data stored in S3. For more information, see Protecting data using server-side encryption with customer-provided encryption keys (SSE-C).

pub fn storage_class(self, input: StorageClass) -> Builder

The class of storage used to store object in Amazon S3.

pub fn set_storage_class(self, input: Option<StorageClass>) -> Builder

The class of storage used to store object in Amazon S3.

pub fn tag_count(self, input: i32) -> Builder

The number of tags, if any, on the object.

pub fn set_tag_count(self, input: Option<i32>) -> Builder

The number of tags, if any, on the object.

pub fn version_id(self, input: impl Into<String>) -> Builder

An ID used to reference a specific version of the object.

pub fn set_version_id(self, input: Option<String>) -> Builder

An ID used to reference a specific version of the object.

pub fn bucket_key_enabled(self, input: bool) -> Builder

Indicates whether the object stored in Amazon S3 uses an S3 bucket key for server-side encryption with Amazon Web Services KMS (SSE-KMS).

pub fn set_bucket_key_enabled(self, input: Option<bool>) -> Builder

Indicates whether the object stored in Amazon S3 uses an S3 bucket key for server-side encryption with Amazon Web Services KMS (SSE-KMS).

pub fn build(self) -> Result<WriteGetObjectResponseInput, BuildError>

Consumes the builder and constructs a WriteGetObjectResponseInput