Struct bonsaidb::keystorage::s3::S3VaultKeyStorage

pub struct S3VaultKeyStorage {
    pub endpoint: Option<Endpoint>,
    pub region: Option<Region>,
    pub path: String,
    /* private fields */
}

S3-compatible VaultKeyStorage implementor.

Fields
endpoint: Option<Endpoint>

The S3 endpoint to use. If not specified, the endpoint will be determined automatically. This field can be used to support non-AWS S3 providers.

region: Option<Region>

The AWS region to use. If not specified, the region will be determined by the aws sdk.

path: String

The path prefix for keys to be stored within.

Implementations

impl S3VaultKeyStorage

pub fn new(bucket: impl Display) -> S3VaultKeyStorage

Creates a new key storage instance for bucket. This instance will use the currently available Tokio runtime or create one if none is available.

pub fn new_with_runtime(
bucket: impl Display,
runtime: Handle
) -> S3VaultKeyStorage

Creates a new key storage instance for bucket, which performs its networking operations on runtime.

pub fn path(self, prefix: impl Display) -> S3VaultKeyStorage

Sets the path prefix for vault keys to be stored within.

pub fn endpoint(self, endpoint: Endpoint) -> S3VaultKeyStorage

Sets the endpoint to use. See Self::endpoint for more information.