Crate bonsaidb::local

BonsaiDb’s offline database implementation.

This crate exposes BonsaiDb’s local database implementation. The Storage type provides its most common functionality by implementing the StorageConnection.

Minimum Supported Rust Version (MSRV)

While this project is alpha, we are actively adopting the current version of Rust. The current minimum version is 1.60.

Modules

argon2

RustCrypto: Argon2

cli

Command-line interface helpers.

config

Configuration options.

vault

Encryption and secret management.

Structs

AsyncDatabase

A database stored in BonsaiDb. This type is designed for use with Tokio. For blocking (non-asynchronous) code, see the Database type instead.

AsyncStorage

A file-based, multi-database, multi-user database engine. This type is designed for use with Tokio. For blocking (non-asynchronous) code, see the Storage type instead.

Database

A database stored in BonsaiDb. This type blocks the current thread when used. See AsyncDatabase for this type’s async counterpart.

Storage

A file-based, multi-database, multi-user database engine. This type blocks the current thread when used. See AsyncStorage for this type’s async counterpart.

StorageId

The unique id of a Storage instance.

Subscriber

A subscriber for PubSub messages.

Enums

Error

Errors that can occur from interacting with storage.

Traits

BackupLocation

A location to store and restore a database from.

DatabaseNonBlocking

Operations that can be performed on both Database and AsyncDatabase.

StorageNonBlocking

Functionality that is available on both Storage and AsyncStorage.