Module bonsaidb::local::core::connection

Types for interacting with BonsaiDb.

Structs

Collection

Interacts with a collection over a Connection.

List

Executes Connection::list() when awaited. Also offers methods to customize the options for the operation.

PasswordResult

The result of logging in with a password or setting a password.

Range

A range type that can represent all std range types and be serialized.

View

Parameters to query a schema::View.

Enums

AccessPolicy

Changes how the view’s outdated data will be treated.

Bound

A range bound that can be serialized.

QueryKey

Filters a View by key.

Sort

A sort order.

Traits

Connection

Defines all interactions with a schema::Schema, regardless of whether it is local or remote.

StorageConnection

Functions for interacting with a multi-database BonsaiDb instance.