Trait bonsaidb::local::core::connection::ServerConnection

pub trait ServerConnection: Send + Sync {

    fn create_database_with_schema<'life0, 'life1, 'async_trait>(
        &'life0 self, 
        name: &'life1 str, 
        schema: SchemaName, 
        only_if_needed: bool
    ) -> Pin<Box<dyn Future<Output = Result<(), Error>> + Send + 'async_trait, Global>>
    where
        'life0: 'async_trait,
        'life1: 'async_trait,
        Self: 'async_trait;

    fn delete_database<'life0, 'life1, 'async_trait>(
        &'life0 self, 
        name: &'life1 str
    ) -> Pin<Box<dyn Future<Output = Result<(), Error>> + Send + 'async_trait, Global>>
    where
        'life0: 'async_trait,
        'life1: 'async_trait,
        Self: 'async_trait;

    fn list_databases<'life0, 'async_trait>(
        &'life0 self
    ) -> Pin<Box<dyn Future<Output = Result<Vec<Database, Global>, Error>> + Send + 'async_trait, Global>>
    where
        'life0: 'async_trait,
        Self: 'async_trait;

    fn list_available_schemas<'life0, 'async_trait>(
        &'life0 self
    ) -> Pin<Box<dyn Future<Output = Result<Vec<SchemaName, Global>, Error>> + Send + 'async_trait, Global>>
    where
        'life0: 'async_trait,
        Self: 'async_trait;

    fn create_user<'life0, 'life1, 'async_trait>(
        &'life0 self, 
        username: &'life1 str
    ) -> Pin<Box<dyn Future<Output = Result<u64, Error>> + Send + 'async_trait, Global>>
    where
        'life0: 'async_trait,
        'life1: 'async_trait,
        Self: 'async_trait;

    fn set_user_password<'user, 'life0, 'async_trait, U>(
        &'life0 self, 
        user: U, 
        password_request: RegistrationRequest
    ) -> Pin<Box<dyn Future<Output = Result<RegistrationResponse, Error>> + Send + 'async_trait, Global>>
    where
        'user: 'async_trait,
        'life0: 'async_trait,
        Self: 'async_trait,
        U: Into<NamedReference<'user>> + Send + Sync + 'async_trait;

    fn finish_set_user_password<'user, 'life0, 'async_trait, U>(
        &'life0 self, 
        user: U, 
        password_finalization: RegistrationFinalization
    ) -> Pin<Box<dyn Future<Output = Result<(), Error>> + Send + 'async_trait, Global>>
    where
        'user: 'async_trait,
        'life0: 'async_trait,
        Self: 'async_trait,
        U: Into<NamedReference<'user>> + Send + Sync + 'async_trait;

    fn add_permission_group_to_user<'user, 'group, 'life0, 'async_trait, U, G>(
        &'life0 self, 
        user: U, 
        permission_group: G
    ) -> Pin<Box<dyn Future<Output = Result<(), Error>> + Send + 'async_trait, Global>>
    where
        'user: 'async_trait,
        'group: 'async_trait,
        'life0: 'async_trait,
        Self: 'async_trait,
        U: Into<NamedReference<'user>> + Send + Sync + 'async_trait,
        G: Into<NamedReference<'group>> + Send + Sync + 'async_trait;

    fn remove_permission_group_from_user<'user, 'group, 'life0, 'async_trait, U, G>(
        &'life0 self, 
        user: U, 
        permission_group: G
    ) -> Pin<Box<dyn Future<Output = Result<(), Error>> + Send + 'async_trait, Global>>
    where
        'user: 'async_trait,
        'group: 'async_trait,
        'life0: 'async_trait,
        Self: 'async_trait,
        U: Into<NamedReference<'user>> + Send + Sync + 'async_trait,
        G: Into<NamedReference<'group>> + Send + Sync + 'async_trait;

    fn add_role_to_user<'user, 'role, 'life0, 'async_trait, U, R>(
        &'life0 self, 
        user: U, 
        role: R
    ) -> Pin<Box<dyn Future<Output = Result<(), Error>> + Send + 'async_trait, Global>>
    where
        'user: 'async_trait,
        'role: 'async_trait,
        'life0: 'async_trait,
        Self: 'async_trait,
        U: Into<NamedReference<'user>> + Send + Sync + 'async_trait,
        R: Into<NamedReference<'role>> + Send + Sync + 'async_trait;

    fn remove_role_from_user<'user, 'role, 'life0, 'async_trait, U, R>(
        &'life0 self, 
        user: U, 
        role: R
    ) -> Pin<Box<dyn Future<Output = Result<(), Error>> + Send + 'async_trait, Global>>
    where
        'user: 'async_trait,
        'role: 'async_trait,
        'life0: 'async_trait,
        Self: 'async_trait,
        U: Into<NamedReference<'user>> + Send + Sync + 'async_trait,
        R: Into<NamedReference<'role>> + Send + Sync + 'async_trait;

    fn create_database<'life0, 'life1, 'async_trait, DB>(
        &'life0 self, 
        name: &'life1 str, 
        only_if_needed: bool
    ) -> Pin<Box<dyn Future<Output = Result<(), Error>> + Send + 'async_trait, Global>>
    where
        'life0: 'async_trait,
        'life1: 'async_trait,
        Self: 'async_trait,
        DB: Schema + 'async_trait,
    { ... }

    fn set_user_password_str<'user, 'life0, 'life1, 'async_trait, U>(
        &'life0 self, 
        user: U, 
        password: &'life1 str
    ) -> Pin<Box<dyn Future<Output = Result<PasswordResult, Error>> + Send + 'async_trait, Global>>
    where
        'user: 'async_trait,
        'life0: 'async_trait,
        'life1: 'async_trait,
        Self: 'async_trait,
        U: Into<NamedReference<'user>> + Send + Sync + 'async_trait,
    { ... }

}

Functions for interacting with a multi-database BonsaiDb instance.

Required methods

fn create_database_with_schema<'life0, 'life1, 'async_trait>(
&'life0 self,
name: &'life1 str,
schema: SchemaName,
only_if_needed: bool
) -> Pin<Box<dyn Future<Output = Result<(), Error>> + Send + 'async_trait, Global>> where
'life0: 'async_trait,
'life1: 'async_trait,
Self: 'async_trait,

Creates a database named name using the SchemaName schema.

Errors

Error::InvalidDatabaseName: name must begin with an alphanumeric character ([a-zA-Z0-9]), and all remaining characters must be alphanumeric, a period (.), or a hyphen (-).

[Error::DatabaseNameAlreadyTaken]: namewas already used for a previous database name. Database names are case insensitive. Returned ifonly_if_needed` is false.

fn delete_database<'life0, 'life1, 'async_trait>(
&'life0 self,
name: &'life1 str
) -> Pin<Box<dyn Future<Output = Result<(), Error>> + Send + 'async_trait, Global>> where
'life0: 'async_trait,
'life1: 'async_trait,
Self: 'async_trait,

Deletes a database named name.

Errors

Error::DatabaseNotFound: database name does not exist.

[Error::Io)]: an error occurred while deleting files.

fn list_databases<'life0, 'async_trait>(
&'life0 self
) -> Pin<Box<dyn Future<Output = Result<Vec<Database, Global>, Error>> + Send + 'async_trait, Global>> where
'life0: 'async_trait,
Self: 'async_trait,

Lists the databases on this server.

fn list_available_schemas<'life0, 'async_trait>(
&'life0 self
) -> Pin<Box<dyn Future<Output = Result<Vec<SchemaName, Global>, Error>> + Send + 'async_trait, Global>> where
'life0: 'async_trait,
Self: 'async_trait,

Lists the SchemaNames on this server.

fn create_user<'life0, 'life1, 'async_trait>(
&'life0 self,
username: &'life1 str
) -> Pin<Box<dyn Future<Output = Result<u64, Error>> + Send + 'async_trait, Global>> where
'life0: 'async_trait,
'life1: 'async_trait,
Self: 'async_trait,

Creates a user.

fn set_user_password<'user, 'life0, 'async_trait, U>(
&'life0 self,
user: U,
password_request: RegistrationRequest
) -> Pin<Box<dyn Future<Output = Result<RegistrationResponse, Error>> + Send + 'async_trait, Global>> where
'user: 'async_trait,
'life0: 'async_trait,
Self: 'async_trait,
U: Into<NamedReference<'user>> + Send + Sync + 'async_trait,

Sets a user’s password using custodian-password to register a password using OPAQUE-PAKE.

fn finish_set_user_password<'user, 'life0, 'async_trait, U>(
&'life0 self,
user: U,
password_finalization: RegistrationFinalization
) -> Pin<Box<dyn Future<Output = Result<(), Error>> + Send + 'async_trait, Global>> where
'user: 'async_trait,
'life0: 'async_trait,
Self: 'async_trait,
U: Into<NamedReference<'user>> + Send + Sync + 'async_trait,

Finishes setting a user’s password by finishing the OPAQUE-PAKE registration.

fn add_permission_group_to_user<'user, 'group, 'life0, 'async_trait, U, G>(
&'life0 self,
user: U,
permission_group: G
) -> Pin<Box<dyn Future<Output = Result<(), Error>> + Send + 'async_trait, Global>> where
'user: 'async_trait,
'group: 'async_trait,
'life0: 'async_trait,
Self: 'async_trait,
U: Into<NamedReference<'user>> + Send + Sync + 'async_trait,
G: Into<NamedReference<'group>> + Send + Sync + 'async_trait,

Adds a user to a permission group.

fn remove_permission_group_from_user<'user, 'group, 'life0, 'async_trait, U, G>(
&'life0 self,
user: U,
permission_group: G
) -> Pin<Box<dyn Future<Output = Result<(), Error>> + Send + 'async_trait, Global>> where
'user: 'async_trait,
'group: 'async_trait,
'life0: 'async_trait,
Self: 'async_trait,
U: Into<NamedReference<'user>> + Send + Sync + 'async_trait,
G: Into<NamedReference<'group>> + Send + Sync + 'async_trait,

Removes a user from a permission group.

fn add_role_to_user<'user, 'role, 'life0, 'async_trait, U, R>(
&'life0 self,
user: U,
role: R
) -> Pin<Box<dyn Future<Output = Result<(), Error>> + Send + 'async_trait, Global>> where
'user: 'async_trait,
'role: 'async_trait,
'life0: 'async_trait,
Self: 'async_trait,
U: Into<NamedReference<'user>> + Send + Sync + 'async_trait,
R: Into<NamedReference<'role>> + Send + Sync + 'async_trait,

Adds a user to a permission group.

fn remove_role_from_user<'user, 'role, 'life0, 'async_trait, U, R>(
&'life0 self,
user: U,
role: R
) -> Pin<Box<dyn Future<Output = Result<(), Error>> + Send + 'async_trait, Global>> where
'user: 'async_trait,
'role: 'async_trait,
'life0: 'async_trait,
Self: 'async_trait,
U: Into<NamedReference<'user>> + Send + Sync + 'async_trait,
R: Into<NamedReference<'role>> + Send + Sync + 'async_trait,

Removes a user from a permission group.

Provided methods

fn create_database<'life0, 'life1, 'async_trait, DB>(
&'life0 self,
name: &'life1 str,
only_if_needed: bool
) -> Pin<Box<dyn Future<Output = Result<(), Error>> + Send + 'async_trait, Global>> where
'life0: 'async_trait,
'life1: 'async_trait,
Self: 'async_trait,
DB: Schema + 'async_trait,

Creates a database named name with the Schema provided.

Errors

Error::InvalidDatabaseName: name must begin with an alphanumeric character ([a-zA-Z0-9]), and all remaining characters must be alphanumeric, a period (.), or a hyphen (-).

[Error::DatabaseNameAlreadyTaken]: namewas already used for a previous database name. Database names are case insensitive. Returned ifonly_if_needed` is false.

fn set_user_password_str<'user, 'life0, 'life1, 'async_trait, U>(
&'life0 self,
user: U,
password: &'life1 str
) -> Pin<Box<dyn Future<Output = Result<PasswordResult, Error>> + Send + 'async_trait, Global>> where
'user: 'async_trait,
'life0: 'async_trait,
'life1: 'async_trait,
Self: 'async_trait,
U: Into<NamedReference<'user>> + Send + Sync + 'async_trait,

Sets a user’s password with the provided string. The password provided will never leave the machine that is calling this function. Internally uses set_user_password and finish_set_user_password in conjunction with custodian-password.