Struct bonsaidb::local::Storage

pub struct Storage { /* private fields */ }

A file-based, multi-database, multi-user database engine. This type blocks the current thread when used. See AsyncStorage for this type’s async counterpart.

Converting between Blocking and Async Types

AsyncStorage and Storage can be converted to and from each other using:

AsyncStorage::into_blocking()

AsyncStorage::to_blocking()

AsyncStorage::as_blocking()

Storage::into_async()

Storage::to_async()

Storage::into_async_with_runtime()

Storage::to_async_with_runtime()

Converting from Database::open to Storage::open

Database::open is a simple method that uses Storage to create a database named default with the schema provided. These two ways of opening the database are the same:

// `bonsaidb_core` is re-exported to `bonsaidb::core` or `bonsaidb_local::core`.
use bonsaidb_core::{connection::StorageConnection, schema::Schema};
// `bonsaidb_local` is re-exported to `bonsaidb::local` if using the omnibus crate.
use bonsaidb_local::{
    config::{Builder, StorageConfiguration},
    Database, Storage,
};
// This creates a Storage instance, creates a database, and returns it.
let db = Database::open::<MySchema>(StorageConfiguration::new("my-db.bonsaidb"))?;

// This is the equivalent code being executed:
let storage =
    Storage::open(StorageConfiguration::new("my-db.bonsaidb").with_schema::<MySchema>()?)?;
storage.create_database::<MySchema>("default", true)?;
let db = storage.database::<MySchema>("default")?;

Using multiple databases

This example shows how to use Storage to create and use multiple databases with multiple schemas:

use bonsaidb_core::{
    connection::StorageConnection,
    schema::{Collection, Schema},
};
use bonsaidb_local::{
    config::{Builder, StorageConfiguration},
    Storage,
};
use serde::{Deserialize, Serialize};

#[derive(Debug, Schema)]
#[schema(name = "my-schema", collections = [BlogPost, Author])]
struct MySchema;

#[derive(Debug, Serialize, Deserialize, Collection)]
#[collection(name = "blog-posts")]
struct BlogPost {
    pub title: String,
    pub contents: String,
    pub author_id: u64,
}

#[derive(Debug, Serialize, Deserialize, Collection)]
#[collection(name = "blog-posts")]
struct Author {
    pub name: String,
}

let storage = Storage::open(
    StorageConfiguration::new("my-db.bonsaidb")
        .with_schema::<BlogPost>()?
        .with_schema::<MySchema>()?,
)?;

storage.create_database::<BlogPost>("ectons-blog", true)?;
let ectons_blog = storage.database::<BlogPost>("ectons-blog")?;
storage.create_database::<MySchema>("another-db", true)?;
let another_db = storage.database::<MySchema>("another-db")?;

Implementations

impl Storage

pub fn backup<L>(&self, location: &L) -> Result<(), Error>where
L: AnyBackupLocation,

Stores a copy of all data in this instance to location.

pub fn restore<L>(&self, location: &L) -> Result<(), Error>where
L: AnyBackupLocation,

Restores all data from a previously stored backup location.

impl Storage

pub fn open(configuration: StorageConfiguration) -> Result<Storage, Error>

Creates or opens a multi-database Storage with its data stored in directory.

pub fn unique_id(&self) -> StorageId

Returns the unique id of the server.

This value is set from the StorageConfiguration or randomly generated when creating a server. It shouldn’t be changed after a server is in use, as doing can cause issues. For example, the vault that manages encrypted storage uses the server ID to store the vault key. If the server ID changes, the vault key storage will need to be updated with the new server ID.

pub fn register_schema<DB>(&self) -> Result<(), Error>where
DB: Schema,

Registers a schema for use within the server.

pub fn with_effective_permissions(
&self,
effective_permissions: Permissions
) -> Option<Storage>

Restricts an unauthenticated instance to having effective_permissions. Returns None if a session has already been established.

pub fn into_async(self) -> AsyncStorage

Converts this instance into its blocking version, which is able to be used without async. The returned instance uses the current Tokio runtime handle to spawn blocking tasks.

Panics

Panics if called outside the context of a Tokio runtime.

pub fn into_async_with_runtime(self, runtime: Handle) -> AsyncStorage

Converts this instance into its blocking version, which is able to be used without async. The returned instance uses the provided runtime handle to spawn blocking tasks.

pub fn to_async(&self) -> AsyncStorage

Converts this instance into its blocking version, which is able to be used without async. The returned instance uses the current Tokio runtime handle to spawn blocking tasks.

Panics

Panics if called outside the context of a Tokio runtime.

pub fn to_async_with_runtime(&self, runtime: Handle) -> AsyncStorage

Converts this instance into its blocking version, which is able to be used without async. The returned instance uses the provided runtime handle to spawn blocking tasks.