Module bonsaidb::server::local::backup

Local database tool to save and load databases into plain an easy-to-consume filesystem structure.

This tool is provided to ensure you always have a way to get your data backed up in a format that’s easily consumable. This tool also provides a safety mechanism allowing a path of migrating across underlying storage layer changes that might have no other migration path.

To back up an existing local database:

bonsaidb local-backup <database_path> save

To restore a backup:

bonsaidb local-backup <database_path> load <backup_location>

Structs

Cli

The command line interface for bonsaidb local-backup.

Enums

Command

The command to execute.