Trait bonsaidb::server::local::core::connection::ServerConnection

pub trait ServerConnection: Send + Sync {
    #[must_use]
    pub fn create_database_with_schema<'life0, 'life1, 'async_trait>(
        &'life0 self, 
        name: &'life1 str, 
        schema: SchemaName
    ) -> Pin<Box<dyn Future<Output = Result<(), Error>> + 'async_trait + Send, Global>>
    where
        'life0: 'async_trait,
        'life1: 'async_trait,
        Self: 'async_trait;

    #[must_use]
    pub fn delete_database<'life0, 'life1, 'async_trait>(
        &'life0 self, 
        name: &'life1 str
    ) -> Pin<Box<dyn Future<Output = Result<(), Error>> + 'async_trait + Send, Global>>
    where
        'life0: 'async_trait,
        'life1: 'async_trait,
        Self: 'async_trait;

    #[must_use]
    pub fn list_databases<'life0, 'async_trait>(
        &'life0 self
    ) -> Pin<Box<dyn Future<Output = Result<Vec<Database, Global>, Error>> + 'async_trait + Send, Global>>
    where
        'life0: 'async_trait,
        Self: 'async_trait;

    #[must_use]
    pub fn list_available_schemas<'life0, 'async_trait>(
        &'life0 self
    ) -> Pin<Box<dyn Future<Output = Result<Vec<SchemaName, Global>, Error>> + 'async_trait + Send, Global>>
    where
        'life0: 'async_trait,
        Self: 'async_trait;

    #[must_use]
    pub fn create_user<'life0, 'life1, 'async_trait>(
        &'life0 self, 
        username: &'life1 str
    ) -> Pin<Box<dyn Future<Output = Result<u64, Error>> + 'async_trait + Send, Global>>
    where
        'life0: 'async_trait,
        'life1: 'async_trait,
        Self: 'async_trait;

    #[must_use]
    pub fn set_user_password<'life0, 'life1, 'async_trait>(
        &'life0 self, 
        username: &'life1 str, 
        password_request: RegistrationRequest
    ) -> Pin<Box<dyn Future<Output = Result<RegistrationResponse, Error>> + 'async_trait + Send, Global>>
    where
        'life0: 'async_trait,
        'life1: 'async_trait,
        Self: 'async_trait;

    #[must_use]
    pub fn finish_set_user_password<'life0, 'life1, 'async_trait>(
        &'life0 self, 
        username: &'life1 str, 
        password_finalization: RegistrationFinalization
    ) -> Pin<Box<dyn Future<Output = Result<(), Error>> + 'async_trait + Send, Global>>
    where
        'life0: 'async_trait,
        'life1: 'async_trait,
        Self: 'async_trait;

    #[must_use]
    pub fn create_database<'life0, 'life1, 'async_trait, DB>(
        &'life0 self, 
        name: &'life1 str
    ) -> Pin<Box<dyn Future<Output = Result<(), Error>> + 'async_trait + Send, Global>>
    where
        'life0: 'async_trait,
        'life1: 'async_trait,
        Self: 'async_trait,
        DB: Schema + 'async_trait,
    { ... }

    #[must_use]
    pub fn set_user_password_str<'life0, 'life1, 'life2, 'async_trait>(
        &'life0 self, 
        username: &'life1 str, 
        password: &'life2 str
    ) -> Pin<Box<dyn Future<Output = Result<ClientFile, Error>> + 'async_trait + Send, Global>>
    where
        'life0: 'async_trait,
        'life1: 'async_trait,
        'life2: 'async_trait,
        Self: 'async_trait,
    { ... }
}

Functions for interacting with a multi-database BonsaiDb instance.

Required methods

#[must_use]
pub fn create_database_with_schema<'life0, 'life1, 'async_trait>(
&'life0 self,
name: &'life1 str,
schema: SchemaName
) -> Pin<Box<dyn Future<Output = Result<(), Error>> + 'async_trait + Send, Global>> where
'life0: 'async_trait,
'life1: 'async_trait,
Self: 'async_trait,

Creates a database named name using the SchemaName schema.

Errors

Error::InvalidDatabaseName: name must begin with an alphanumeric character ([a-zA-Z0-9]), and all remaining characters must be alphanumeric, a period (.), or a hyphen (-).

[Error::DatabaseNameAlreadyTaken]: name` was already used for a previous database name. Database names are case insensitive.

#[must_use]
pub fn delete_database<'life0, 'life1, 'async_trait>(
&'life0 self,
name: &'life1 str
) -> Pin<Box<dyn Future<Output = Result<(), Error>> + 'async_trait + Send, Global>> where
'life0: 'async_trait,
'life1: 'async_trait,
Self: 'async_trait,

Deletes a database named name.

Errors

Error::DatabaseNotFound: database name does not exist.

[Error::Io)]: an error occurred while deleting files.

#[must_use]
pub fn list_databases<'life0, 'async_trait>(
&'life0 self
) -> Pin<Box<dyn Future<Output = Result<Vec<Database, Global>, Error>> + 'async_trait + Send, Global>> where
'life0: 'async_trait,
Self: 'async_trait,

Lists the databases on this server.

#[must_use]
pub fn list_available_schemas<'life0, 'async_trait>(
&'life0 self
) -> Pin<Box<dyn Future<Output = Result<Vec<SchemaName, Global>, Error>> + 'async_trait + Send, Global>> where
'life0: 'async_trait,
Self: 'async_trait,

Lists the SchemaNames on this server.

#[must_use]
pub fn create_user<'life0, 'life1, 'async_trait>(
&'life0 self,
username: &'life1 str
) -> Pin<Box<dyn Future<Output = Result<u64, Error>> + 'async_trait + Send, Global>> where
'life0: 'async_trait,
'life1: 'async_trait,
Self: 'async_trait,

Creates a user.

#[must_use]
pub fn set_user_password<'life0, 'life1, 'async_trait>(
&'life0 self,
username: &'life1 str,
password_request: RegistrationRequest
) -> Pin<Box<dyn Future<Output = Result<RegistrationResponse, Error>> + 'async_trait + Send, Global>> where
'life0: 'async_trait,
'life1: 'async_trait,
Self: 'async_trait,

Sets a user’s password using custodian-password to register a password using OPAQUE-PAKE.

#[must_use]
pub fn finish_set_user_password<'life0, 'life1, 'async_trait>(
&'life0 self,
username: &'life1 str,
password_finalization: RegistrationFinalization
) -> Pin<Box<dyn Future<Output = Result<(), Error>> + 'async_trait + Send, Global>> where
'life0: 'async_trait,
'life1: 'async_trait,
Self: 'async_trait,

Finishes setting a user’s password by finishing the OPAQUE-PAKE registration.

Loading content...
Provided methods

#[must_use]
pub fn create_database<'life0, 'life1, 'async_trait, DB>(
&'life0 self,
name: &'life1 str
) -> Pin<Box<dyn Future<Output = Result<(), Error>> + 'async_trait + Send, Global>> where
'life0: 'async_trait,
'life1: 'async_trait,
Self: 'async_trait,
DB: Schema + 'async_trait,

Creates a database named name with the Schema provided.

Errors

Error::InvalidDatabaseName: name must begin with an alphanumeric character ([a-zA-Z0-9]), and all remaining characters must be alphanumeric, a period (.), or a hyphen (-).

[Error::DatabaseNameAlreadyTaken]: name` was already used for a previous database name. Database names are case insensitive.

#[must_use]
pub fn set_user_password_str<'life0, 'life1, 'life2, 'async_trait>(
&'life0 self,
username: &'life1 str,
password: &'life2 str
) -> Pin<Box<dyn Future<Output = Result<ClientFile, Error>> + 'async_trait + Send, Global>> where
'life0: 'async_trait,
'life1: 'async_trait,
'life2: 'async_trait,
Self: 'async_trait,

Sets a user’s password with the provided string. The password provided will never leave the machine that is calling this function. Internally uses set_user_password and finish_set_user_password in conjunction with custodian-password.

Loading content...