Trait bonsaidb::server::local::core::schema::view::map::Key

pub trait Key: Clone + Send + Sync {
    type Error: AnyError;
    fn as_big_endian_bytes(&self) -> Result<Cow<'_, [u8]>, Self::Error>;

    fn from_big_endian_bytes(bytes: &[u8]) -> Result<Self, Self::Error>;
}

A trait that enables a type to convert itself to a big-endian/network byte order.

Associated Types

type Error: AnyError

The error type that can be produced by either serialization or deserialization.

Required methods

fn as_big_endian_bytes(&self) -> Result<Cow<'_, [u8]>, Self::Error>

Convert self into a Cow<[u8]> containing bytes ordered in big-endian/network byte order.

fn from_big_endian_bytes(bytes: &[u8]) -> Result<Self, Self::Error>

Convert a slice of bytes into Self by interpretting bytes in big-endian/network byte order.