Struct bonsaidb::server::CustomServer

pub struct CustomServer<B> where
    B: Backend,  { /* fields omitted */ }

A BonsaiDb server.

Implementations

impl<B> CustomServer<B> where
B: Backend,

pub async fn open(
directory: &'_ Path,
configuration: Configuration
) -> Result<CustomServer<B>, Error>

Opens a server using directory for storage.

pub fn directory(&self) -> &PathBuf

Returns the path to the directory that stores this server’s data.

pub async fn database<DB>(
&'_ self,
name: &'_ str
) -> Result<ServerDatabase<'_, B, DB>, Error> where
DB: Schema,

Retrieves a database. This function only verifies that the database exists.

pub async fn admin(&'_ self) -> ServerDatabase<'_, B, Admin>

Returns the administration database.

pub async fn install_self_signed_certificate(
&'_ self,
server_name: &'_ str,
overwrite: bool
) -> Result<(), Error>

Installs an X.509 certificate used for general purpose connections.

pub async fn install_certificate(
&'_ self,
certificate: &'_ Certificate,
private_key: &'_ PrivateKey
) -> Result<(), Error>

Installs an X.509 certificate used for general purpose connections. These currently must be in DER binary format, not ASCII PEM format.

pub async fn certificate(&'_ self) -> Result<Certificate, Error>

Returns the current certificate.

pub async fn listen_on(&'_ self, port: u16) -> Result<(), Error>

Listens for incoming client connections. Does not return until the server shuts down.

pub async fn listen_for_websockets_on<T>(&'_ self, addr: T) -> Result<(), Error> where
T: ToSocketAddrs + Send + Sync,

Listens for WebSocket traffic on port.

pub async fn connected_clients(&'_ self) -> Vec<ConnectedClient<B>, Global>

Returns all of the currently connected clients.

pub async fn broadcast(
&'_ self,
response: <<B as Backend>::CustomApi as CustomApi>::Response
)

Sends a custom API response to all connected clients.

pub async fn shutdown(&'_ self, timeout: Option<Duration>) -> Result<(), Error>

Shuts the server down. If a timeout is provided, the server will stop accepting new connections and attempt to respond to any outstanding requests already being processed. After the timeout has elapsed or if no timeout was provided, the server is forcefully shut down.