Struct bonsaidb::server::ServerConfiguration

#[non_exhaustive]
pub struct ServerConfiguration {
    pub server_name: String,
    pub client_simultaneous_request_limit: usize,
    pub request_workers: usize,
    pub storage: StorageConfiguration,
    pub default_permissions: DefaultPermissions,
    pub authenticated_permissions: DefaultPermissions,
    pub acme: AcmeConfiguration,
}

Configuration options for Server

Fields (Non-exhaustive)

Non-exhaustive structs could have additional fields added in future. Therefore, non-exhaustive structs cannot be constructed in external crates using the traditional Struct { .. } syntax; cannot be matched against without a wildcard ..; and struct update syntax will not work.

server_name: String

The DNS name of the server.

client_simultaneous_request_limit: usize

Number of sumultaneous requests a single client can have in flight at a time. Default value is 16. It is important to have this number be tuned relative to request_workers such that one client cannot overwhelm the entire queue.

request_workers: usize

Number of simultaneous requests to be processed. Default value is 16.

storage: StorageConfiguration

Configuration options for individual databases.

default_permissions: DefaultPermissions

The permissions granted to all connections to this server.

authenticated_permissions: DefaultPermissions

The permissions granted to authenticated connections to this server.

acme: AcmeConfiguration

The ACME settings for automatic TLS certificate management.

Implementations

impl ServerConfiguration

pub fn server_name(self, server_name: impl Into<String>) -> ServerConfiguration

Sets Self::server_name to server_name and returns self.

pub const fn client_simultaneous_request_limit(
self,
request_limit: usize
) -> ServerConfiguration

Sets Self::client_simultaneous_request_limit to request_limit and returns self.

pub const fn request_workers(self, workers: usize) -> ServerConfiguration

Sets Self::request_workers to workers and returns self.

pub fn default_permissions<P>(
self,
default_permissions: P
) -> ServerConfiguration where
P: Into<DefaultPermissions>,

Sets Self::default_permissions to default_permissions and returns self.

pub fn authenticated_permissions<P>(
self,
authenticated_permissions: P
) -> ServerConfiguration where
P: Into<DefaultPermissions>,

Sets Self::authenticated_permissions to authenticated_permissions and returns self.

pub fn acme_contact_email(
self,
contact_email: impl Into<String>
) -> ServerConfiguration

Sets AcmeConfiguration::contact_email to contact_email and returns self.

pub fn acme_directory(self, directory: impl Into<String>) -> ServerConfiguration

Sets AcmeConfiguration::directory to directory and returns self.