Trait bonsaidb::server::Backend

pub trait Backend: 'static + Debug + Send + Sync {
    type CustomApi: CustomApi;
    type CustomApiDispatcher: Dispatcher<<Self::CustomApi as CustomApi>::Request> + Debug;
    pub fn dispatcher_for(
        server: &CustomServer<Self>, 
        client: &ConnectedClient<Self>
    ) -> Self::CustomApiDispatcher;

    #[must_use]
    #[must_use]
    pub fn client_connected<'life0, 'life1, 'async_trait>(
        client: &'life0 ConnectedClient<Self>, 
        server: &'life1 CustomServer<Self>
    ) -> Pin<Box<dyn Future<Output = ConnectionHandling> + 'async_trait + Send, Global>>
    where
        'life0: 'async_trait,
        'life1: 'async_trait,
        Self: 'async_trait,
    { ... }

    #[must_use]
    pub fn client_disconnected<'life0, 'async_trait>(
        client: ConnectedClient<Self>, 
        server: &'life0 CustomServer<Self>
    ) -> Pin<Box<dyn Future<Output = ()> + 'async_trait + Send, Global>>
    where
        'life0: 'async_trait,
        Self: 'async_trait,
    { ... }

    #[must_use]
    pub fn client_authenticated<'life0, 'async_trait>(
        client: ConnectedClient<Self>, 
        server: &'life0 CustomServer<Self>
    ) -> Pin<Box<dyn Future<Output = ()> + 'async_trait + Send, Global>>
    where
        'life0: 'async_trait,
        Self: 'async_trait,
    { ... }
}

Tailors the behavior of a server to your needs.

Associated Types

type CustomApi: CustomApi

The custom API definition. If you do not wish to have an API, () may be provided.

type CustomApiDispatcher: Dispatcher<<Self::CustomApi as CustomApi>::Request> + Debug

The type that implements the Dispatcher trait.

Loading content...
Required methods

pub fn dispatcher_for(
server: &CustomServer<Self>,
client: &ConnectedClient<Self>
) -> Self::CustomApiDispatcher

Returns a dispatcher to handle custom api requests. The server and client parameters are provided to allow the dispatcher to have access to them when handling the individual actions.

Loading content...
Provided methods

#[must_use]

#[must_use]
pub fn client_connected<'life0, 'life1, 'async_trait>(
client: &'life0 ConnectedClient<Self>,
server: &'life1 CustomServer<Self>
) -> Pin<Box<dyn Future<Output = ConnectionHandling> + 'async_trait + Send, Global>> where
'life0: 'async_trait,
'life1: 'async_trait,
Self: 'async_trait,

A client disconnected from the server. This is invoked before authentication has been performed.

#[must_use]
pub fn client_disconnected<'life0, 'async_trait>(
client: ConnectedClient<Self>,
server: &'life0 CustomServer<Self>
) -> Pin<Box<dyn Future<Output = ()> + 'async_trait + Send, Global>> where
'life0: 'async_trait,
Self: 'async_trait,

A client disconnected from the server.

#[must_use]
pub fn client_authenticated<'life0, 'async_trait>(
client: ConnectedClient<Self>,
server: &'life0 CustomServer<Self>
) -> Pin<Box<dyn Future<Output = ()> + 'async_trait + Send, Global>> where
'life0: 'async_trait,
Self: 'async_trait,

A client successfully authenticated.

Loading content...