Module bonsaidb::core::document

Types for interacting with Documents.

Structs

BorrowedDocument

Contains a serialized document in the database.

CollectionDocument

A document with serializable contents.

CollectionHeader

A header for a CollectionDocument.

DocumentId

The serialized representation of a document’s unique ID.

Header

The header of a Document.

InvalidHexadecimal

An invalid hexadecimal character was encountered.

OwnedDocument

Contains a serialized document in the database.

Revision

Information about a Document’s revision history.

Enums

AnyDocumentId

A unique id for a document, either serialized or deserialized.

AnyHeader

A header with either a serialized or deserialized primary key.

KeyId

The ID of an encryption key.

Traits

Document

Common interface of a document in BonsaiDb.

Emit

View mapping emit functions. Used when implementing a view’s map() function.

HasHeader

A type that can return a Header.

OwnedDocuments

Helper functions for a slice of OwnedDocuments.