Trait bonsaidb::core::key::Key

pub trait Key<'k>: Clone + Debug + Send + Sync {
    type Error: AnyError;

    const LENGTH: Option<usize>;

    fn as_ord_bytes(&'k self) -> Result<Cow<'k, [u8]>, Self::Error>;

    fn from_ord_bytes(bytes: &'k [u8]) -> Result<Self, Self::Error>;

    fn first_value() -> Result<Self, NextValueError> { ... }

    fn next_value(&self) -> Result<Self, NextValueError> { ... }
}

A trait that enables a type to convert itself into a memcmp-compatible sequence of bytes.

Associated Types

type Error: AnyError

The error type that can be produced by either serialization or deserialization.

Associated Constants

const LENGTH: Option<usize>

The size of the key, if constant. If this type doesn’t produce the same number of bytes for each value, this should be None.

Required methods

fn as_ord_bytes(&'k self) -> Result<Cow<'k, [u8]>, Self::Error>

Convert self into a Cow<[u8]> containing bytes that are able to be compared via memcmp in a way that is comptaible with its own Ord implementation.

fn from_ord_bytes(bytes: &'k [u8]) -> Result<Self, Self::Error>

Deserialize a sequence of bytes previously encoded with Self::as_ord_bytes.

Provided methods

fn first_value() -> Result<Self, NextValueError>

Return the first value in sequence for this type. Not all types implement this.

fn next_value(&self) -> Result<Self, NextValueError>

Return the next value in sequence for this type. Not all types implement this. Instead of wrapping/overflowing, None should be returned.