Module bonsaidb::core::permissions::bonsai

Types used for granting permissions within BonsaiDb.

Enums

BonsaiAction

Actions that can be permitted within BonsaiDb.

DatabaseAction

Actions that operate on a specific database.

DocumentAction

Actions that operate on a document.

EncryptionKeyAction

Actions that use encryption keys.

KeyValueAction

Actions that operate on the key-value store.

PubSubAction

Actions that operate on the PubSub system.

ServerAction

Actions that operate on a server.

TransactionAction

Actions that operate on transactions.

ViewAction

Actions that operate on a view.

Functions

authentication_token_resource_name

Creates a resource name for token_id.

bonsaidb_resource_name

The base BonsaiDb resource namespace. All database objects have this as their first name segment.

collection_resource_name

Creates a resource name for a collection within a database.

database_resource_name

Creates a resource name with the database name.

document_resource_name

Creates a resource name for a document id within collection within database.

encryption_key_resource_name

Creates a resource name for encryption key key_id.

keyvalue_key_resource_name

Creates a resource name for key within namespace within the key-value store of database.

kv_resource_name

Creates a resource name for the key-value store in database.

pubsub_topic_resource_name

Creates a resource name for PubSub topic within database.

role_resource_name

Creates a resource name for role_id.

user_resource_name

Creates a resource name for user_id.

view_resource_name

Creaets a resource name for a view within database.