Module bonsaidb::core::schema::view

Types for defining map/reduce-powered Views.

Modules

map

Types for defining a Map within a View.

Macros

define_basic_mapped_view

Defines a non-unique view named $view_name for $collection with the mapping provided.

define_basic_unique_mapped_view

Defines an unique view named $view_name for $collection with the mapping provided.

define_mapped_view

Defines a view using the mapping provided.

Enums

Error

Errors that arise when interacting with views.

Traits

CollectionViewSchema

A View for a Collection that stores Serde-compatible documents. The only difference between implmementing this and View is that the map function receives a CollectionDocument instead of a BorrowedDocument.

DefaultViewSerialization

A default serialization strategy for views. Uses equivalent settings as DefaultSerialization.

Serialized

Wraps a View with serialization to erase the associated types

SerializedView

A View with additional tyes and logic to handle serializing view values.

View

An lazy index of mapped and/or reduced data from a Collection.

ViewSchema

The implementation of Map/Reduce for a View.

Type Definitions

ReduceResult

A type alias for the result of ViewSchema::reduce().

ViewMapResult

A type alias for the result of ViewSchema::map().