Struct bonsaidb::server::CustomServer

pub struct CustomServer<B = NoBackend> where
    B: Backend,  { /* fields omitted */ }

A BonsaiDb server.

Implementations

impl<B> CustomServer<B> where
B: Backend,

pub async fn listen_for_tcp_on<S, T>(
&'_ self,
addr: T,
service: S
) -> Result<(), Error> where
S: TcpService,
T: ToSocketAddrs + Send + Sync,

Listens for HTTP traffic on port. This port will also receive WebSocket connections if feature websockets is enabled.

pub async fn listen_for_secure_tcp_on<S, T>(
&'_ self,
addr: T,
service: S
) -> Result<(), Error> where
S: TcpService,
T: ToSocketAddrs + Send + Sync,

Listens for HTTPS traffic on port. This port will also receive WebSocket connections if feature websockets is enabled. If feature acme is enabled, this connection will automatically manage the server’s private key and certificate, which is also used for the QUIC-based protocol.

impl<B> CustomServer<B> where
B: Backend,

pub async fn listen_for_websockets_on<T>(
&'_ self,
addr: T,
with_tls: bool
) -> Result<(), Error> where
T: ToSocketAddrs + Send + Sync,

Listens for websocket connections on addr.

pub async fn upgrade_websocket(
&'_ self,
peer_address: SocketAddr,
request: Request<Body>
) -> Response<Body>

Handles upgrading an HTTP connection to the WebSocket protocol based on the upgrade request. Requires feature hyper to be enabled.

pub async fn handle_websocket<S, E>(
&'_ self,
connection: S,
peer_address: SocketAddr
) where
S: 'static + Stream<Item = Result<Message, E>> + Sink<Message> + Send,
E: Debug + Send,

Handles an established tokio-tungstenite WebSocket stream.

impl<B> CustomServer<B> where
B: Backend,

pub async fn open(
configuration: ServerConfiguration
) -> Result<CustomServer<B>, Error>

Opens a server using directory for storage.

pub fn pinned_certificate_path(&self) -> PathBuf

Returns the path to the public pinned certificate, if this server has one. Note: this function will always succeed, but the file may not exist.

pub fn primary_domain(&self) -> &str

Returns the primary domain configured for this server.

pub async fn admin(&'_ self) -> ServerDatabase<B>

Returns the administration database.

pub async fn install_self_signed_certificate(
&'_ self,
overwrite: bool
) -> Result<(), Error>

Installs an X.509 certificate used for general purpose connections.

pub async fn install_pem_certificate(
&'_ self,
certificate_chain: &'_ [u8],
private_key: &'_ [u8]
) -> Result<(), Error>

Installs a certificate chain and private key used for TLS connections.

pub async fn install_certificate(
&'_ self,
certificate_chain: &'_ CertificateChain,
private_key: &'_ PrivateKey
) -> Result<(), Error>

Installs a certificate chain and private key used for TLS connections.

pub async fn certificate_chain(&'_ self) -> Result<CertificateChain, Error>

Returns the current certificate chain.

pub async fn listen_on(&'_ self, port: u16) -> Result<(), Error>

Listens for incoming client connections. Does not return until the server shuts down.

pub async fn connected_clients(&'_ self) -> Vec<ConnectedClient<B>, Global>

Returns all of the currently connected clients.

pub async fn broadcast(
&'_ self,
response: Result<<<B as Backend>::CustomApi as CustomApi>::Response, <<B as Backend>::CustomApi as CustomApi>::Error>
)

Sends a custom API response to all connected clients.

pub async fn shutdown(&'_ self, timeout: Option<Duration>) -> Result<(), Error>

Shuts the server down. If a timeout is provided, the server will stop accepting new connections and attempt to respond to any outstanding requests already being processed. After the timeout has elapsed or if no timeout was provided, the server is forcefully shut down.

pub async fn listen_for_shutdown(&'_ self) -> Result<(), Error>

Listens for signals from the operating system that the server should shut down and attempts to gracefully shut down.

pub async fn authenticate_client_as<'name, N>(
&'_ self,
user: N,
client: &'_ ConnectedClient<B>
) -> Result<Permissions, Error> where
N: Into<NamedReference<'name>> + Send + Sync,

Manually authenticates client as user. user can be the user’s id (u64) or the username (String/str). Returns the permissions that the user now has.